% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tina_align_max.R
\name{align.max}
\alias{align.max}
\title{Aligns the features in the input matrix based on their maximum peaks}
\usage{
align.max(feature, scaling = TRUE)
}
\arguments{
\item{feature}{a list object containing a matrix of feature values and a matrix of their corresponding positions}

\item{scaling}{a logical value indicating whether to scale the feature values to between 0 and 1 based on the highest peak's intensity}
}
\value{
a list object containing
}
\description{
Aligns the features in the input matrix based on their maximum peaks
}
