% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/align.spec2feat.R
\name{align.spec2feat}
\alias{align.spec2feat}
\title{Calculate fft-based alignment of spectrum to feature.
Report lag table.}
\usage{
align.spec2feat(feat, xmat, r.thresh = 0.8)
}
\arguments{
\item{feat}{feature object for single feature (with profile, position, and subset vects, and driver position)}

\item{xmat}{spectral matrix}

\item{r.thresh}{correlation threshold}
}
\value{
lag table for each spectrum in xmat
}
\description{
Calculate fft-based alignment of spectrum to feature.
Report lag table.
}
\author{
MTJ
}
