% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_matching.R
\name{annotation_matching}
\alias{annotation_matching}
\title{Annotation Module}
\usage{
annotation_matching(peaks, target, ppeaks, spec, params, output_dir)
}
\arguments{
\item{peaks}{initial peaks picked (driver for picked peaks and target)}

\item{target}{full resolution STOCSY result. One element per STOCSY cluster,
containing ppm vector and thresholded intensity for each ppm value}

\item{ppeaks}{peak picked STOCYS result. One element per STOCSY cluster,
containing a numeric vector: col1: (named "chemical shift") ppm values,
col2: (named "intensity") intensity (covariance from STOCSY).}

\item{spec}{spectral data (also ppm vector).#     File should contain:
 List containing compound entries from hmdb or other db, SORTED by hmdb ID:
    spec_ID
      > metadata
        [various fields]
      > spectrum_peaks
        double (n x 2 array = n peaks x (ppm + intensity in STOCSY))
      > biospecimen
        (char array; e.g.
"Blood" "Cerebrospinal Fluid (CSF)" "Feces" "Urine" )
      > dsource
        (char, data source e.g. 'spec_file' or 'metabocard')}
}
\value{
nothing, export matches is called.
}
\description{
Given picked peaks from statistical decomposition module, do a simple
comparison between each query chemical shift list (STOCSY-derived)
and each hmdb reference spectrum chemical shift list. Optimize the
matching using the Hungarian Algorithm for linear assignment, after
all pairs with distances exceeding the threshold are removed.
}
