% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.fit.R
\name{apply.fit}
\alias{apply.fit}
\title{Helper function to take fit coeffs in match.info, with feat and ref stacks,
to expand to the fit features and spec region.}
\usage{
apply.fit(mi.row, feat.stack, ref.stack)
}
\arguments{
\item{mi.row}{match.info row (mi.row)}

\item{feat.stack}{feature matrix, e.g. feature$stack (feats are rows)}

\item{ref.stack}{ref.mat (refs are cols)}
}
\value{
fit feature and spec (ref) regions
}
\description{
Helper function to take fit coeffs in match.info, with feat and ref stacks,
to expand to the fit features and spec region.
}
\author{
MTJ
}
