% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply.lags.feat.R
\name{apply.lags.feat}
\alias{apply.lags.feat}
\title{Given lags and feature object for single feature (including positions and subset),
make a lagged spectral matrix that is aligned to the feature in its region.}
\usage{
apply.lags.feat(feat, xmat, lags)
}
\arguments{
\item{feat}{feature object for single feature (including positions and subset)}

\item{xmat}{spectral matrix}

\item{lags}{vector of lag values (feat$position + lag gives spec pos)}
}
\value{
list with lagged spectral region indices, values, subset, and range of columns in xmat
}
\description{
Given lags and feature object for single feature (including positions and subset),
make a lagged spectral matrix that is aligned to the feature in its region.
}
\author{
MTJ
}
