% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/apply_fit.R
\name{apply_fit}
\alias{apply_fit}
\title{Helper function to take fit coeffs in match.info, with feat and ref stacks,
to expand to the fit features and spec region.}
\usage{
apply_fit(mi.row, feat.cstack, ref.cstack)
}
\arguments{
\item{mi.row}{match.info row (mi.row)}

\item{feat.cstack}{feature matrix, e.g. feature$stack (feats are rows), compressed using compress_stack()}

\item{ref.cstack}{ref.mat (refs are cols), compressed using compress_stack()}
}
\value{
fit feature and spec (ref) regions
}
\description{
Helper function to take fit coeffs in match.info, with feat and ref stacks,
to expand to the fit features and spec region.
}
\author{
MTJ
}
