% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backfit_rf_2_spec_2.R
\name{backfit_rfs}
\alias{backfit_rfs}
\title{Back-fit reference features to a subset of spectra}
\usage{
backfit_rfs(match.info, feature, xmat, ref.mat, ncores)
}
\arguments{
\item{match.info}{A data frame containing information about the matches between features and reference spectra}

\item{feature}{List containing feature intensities and positions}

\item{xmat}{Full spectral matrix from which the features were derived}
}
\value{
A list containing the back-fits and back-fit feasibility scores for each spectrum in the subset,
        as well as a grid plot of the fits and back-fit feasibility scores.
}
\description{
This has gone through several iterations.
A list of backfits (full fit objects) to each subset spectrum was kept
for each ref feature. The backfit object was really all that was needed to 
estimate a score.
}
\details{
Then, I moved to slimming them down by cutting out data - just storing the fit
coefficients instead of fit rfs explicitly. This was a great reduction. However, 
there was still massive data duplication (e.g. ref, feature, match inds apply 
to all subset spectra - nspec times!).

So, currently, the match.info contains the data that applies at the rf level - 
feat ind, ref ind, rmse (feature), feat region, refspec region - this would all
have been duplicated. The backfits are a list of nrow(match.info) data frames,
each storing the ss-level information (fit coeffs, ss ind, bff scores) for that
rf. The combination of these two tables contains a truly ridiculous amount of
information, as they describe all the shape relationships between the PRCSs and 
dataset spectra. 

The match.info table can be used to filter for rfs which apply to a given
PRCS and/or spectral region, or to a given feature. Once the row (rf) inds are 
obtained, one can easily access the ss-level information for each of the rfs.
This can be used, in combination with xmat and refmat to extract the shapes,
and fits can be applied for plotting, scoring, etc.. Example:
  # 1: select ref
   - this narrows down matches without ref
   - apply to backfits (indexed by match)
   2: select ref region
     - this narrows down matches further
     - still don't need anything from the backfits
     3: select samples
       - now backfits needed
       - plotting can take place. Access:
        - spec region (really just the start)
         - ref region
         - ref spec
         - fit data

Alternatively, one can loop through the rfs relevant to a given PRCS, then pull
the ss-level table for each, and use that for scoring.
}
