% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bad.peaks.R
\name{bad.peaks}
\alias{bad.peaks}
\title{Calculate the probability that a peak is "bad"}
\usage{
bad.peaks(pos.res.pct.feat, scale.exp = 2)
}
\arguments{
\item{pos.res.pct.feat}{matrix of positional residue differences between a 
reference and a feature spectrum, as output by \code{positionalResidue}.}

\item{scale.exp}{a scaling exponent for the probability values (default is 2)}
}
\value{
A vector of probabilities, one for each peak in the feature spectrum,
  indicating the probability that the peak is "bad".
}
\description{
Detect likelihood of each point being a bad peak
mean of the % of the feature signal that had no
corresponding ref signal (extra peak in feature)
useful to scale this by some exponential (e.g. 2) to select extremely high percentages
(higher values does less squashing of the bad peaks)
opf.mat <- lapply(allmatches.fits, function(x) x$overshoot.pct.feat) %>% do.call(rbind,.)
output is a vector which can be (e.g.) multiplied by and then subtracted from fit$fit.feat
}
