% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batman_local_opt.R
\name{batman_fit_in}
\alias{batman_fit_in}
\title{Don't optimize, just fit}
\usage{
batman_fit_in(feat, spec.segment, window.pos, exclude.lowest = 0.5)
}
\arguments{
\item{sf}{spec feature information (includes backfit information)}

\item{feature}{features object for study}

\item{xmat}{spectral matrix for study}

\item{refmat}{reference spectral matrix for study (spectra on rows)}
}
\value{
list: fit optimal fit information
              lag optimal lag value
              fit.ref (same size as spec.segment)
              fit.spec (same size as spec.segment)
}
\description{
Don't optimize, just fit
}
