% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batman_local_opt.R
\name{batman_local_opt}
\alias{batman_local_opt}
\title{Batman-fit a shorter profile (feat) to a longer segment (spec.segment) at 
each internal lag. Pick the optimal lag based on the fraction.spec.accounted
by the fit (where the two signals overlap, excluding NAs). This works because
the Batman fitting does not allow feature signal to exceed spec.segment intensity
(excluding the exclude.lowest fraction of points in feature). The rationale is
that a feature which matches well will account for more of the relevant spec 
signal once fit. Since batman fitting is just calculation of ratios, this goes
pretty quick.}
\usage{
batman_local_opt(
  feat,
  spec.segment,
  window.pos,
  exclude.lowest = 0.5,
  opt.on = "fraction.spec.accounted"
)
}
\arguments{
\item{sf}{spec feature information (includes backfit information)}

\item{feature}{features object for study}

\item{xmat}{spectral matrix for study}

\item{refmat}{reference spectral matrix for study (spectra on rows)}
}
\value{
list: fit optimal fit information
              lag optimal lag value
              fit.ref (same size as spec.segment)
              fit.spec (same size as spec.segment)
}
\description{
Used to optimize after convolution-based matching during matching, or during 
backfitting.
}
