% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse_evidence.R
\name{browse_evidence}
\alias{browse_evidence}
\title{browse_evidence : results viewer shiny app}
\usage{
browse_evidence(
  results.dir = NULL,
  select.compounds = NULL,
  select.samples = NULL
)
}
\arguments{
\item{results.dir}{A list of input parameters.}
}
\value{
spins up a shiny app from the results directory
}
\description{
Browse reference spectra (PRCSs) and metabolite-sample summed scores, and drill
down to the evidence for their associations at the multiple levels.
}
\details{
To use, call with the results directory.
}
