% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.cluster.R
\name{check.cluster}
\alias{check.cluster}
\title{Calculate feature cluster information
- align features
- choose representative (key) feature
- calculate profile for cluster}
\usage{
check.cluster(clust.feats, feature.stack)
}
\arguments{
\item{clust.feats}{cluster feature indices (rows in feature.stack)}

\item{feature.stack}{matrix holding feature profiles (feature$stack)}
}
\value{
clust.info list; labels of features, rmses, profile, key feature, lag.table, other info
}
\description{
Calculate feature cluster information
- align features
- choose representative (key) feature
- calculate profile for cluster
}
