% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_heatmap_html_output.R
\name{check_heatmap_html_output}
\alias{check_heatmap_html_output}
\title{check_heatmap_html_output}
\usage{
check_heatmap_html_output(filename, copy.to = NULL)
}
\arguments{
\item{copy.to}{(optional) full filepath location where you want the html file copied to}

\item{pars}{a list containing necessary parameters for the function}
}
\value{
RDS file containing match scores between library reference spectra and the best subset spectrum score for each.
}
\description{
Check the provided results file for correct standalone html heatmap output.
}
\details{
If file is a zip, unzip first to a temp directory, replace the file, 
then re-zip and remove the temp directory. 

If file is a directory, simply re-write the html heatmap file as a standalone.
}
