% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusts2labs.R
\name{clusts2labs}
\alias{clusts2labs}
\title{Convert clusters to feature-sorted cluster labels}
\usage{
clusts2labs(clusters)
}
\arguments{
\item{clusters}{A list of cluster-feature pairs.}
}
\value{
A numeric vector of feature-sorted cluster labels.
}
\description{
Given a list of cluster-feature pairs, this function returns feature-sorted 
cluster labels.
}
\author{
MTJ
}
