% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{co_compress}
\alias{co_compress}
\title{Sparse (NA-gapped) matrix compression from one matrix applied to others}
\usage{
co_compress(
  stack.list,
  sparse.val = NA,
  key = 1,
  apply.to = NULL,
  stack.names = NULL
)
}
\arguments{
\item{stack.list}{list of matrices with matching positions (same sizes)}

\item{sparse.val}{passthrough to compress_stack()}

\item{key}{which matrix (name or list index) to use as the compression model (default = 1). If name, will be converted to list index.}

\item{apply.to}{which matrices (other than key) to apply compression scheme to}
}
\value{
compressed matrix (list format)
         - pos : linear index within stack
         - vals: non-NA vals within stack
         - m : number of rows in stack
         - n : number of columns in stack
}
\description{
Generalized case of stack.lists matrix
}
\details{
Example: stack <- rbind(c(NA, NA, 1, 3, 5, 3, NA, 1, NA), c(NA, 1, 3, 5, NA, 3, NA, 1, 3), c(3, 5, NA, NA, 1, NA, NA, 1, 3))
         stack.list <- list(a = stack, b = stack-1)
         cstack <- co_compress(stack.list, sparse.val = NA, key = 'a')
}
