% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{compress_features}
\alias{compress_features}
\title{Sparse (NA-gapped) matrix compression applied to features object.}
\usage{
compress_features(feature)
}
\arguments{
\item{feature}{feature object with 
- stack (profiles on rows)
- position (cols in xmat on rows), same size as stack
- driver.relative and sfe optional}
}
\value{
compressed matrix (list format)
         - pos : linear index within stack
         - vals: non-NA vals within stack
         - m : number of rows in stack
         - n : number of columns in stack
}
\description{
In the case of feature objects, the profile and position mats can be 
compressed together.
}
\details{
Example: stack <- rbind(c(NA, NA, 1, 3, 5, 3, NA, 1, NA), c(NA, 1, 3, 5, NA, 3, NA, 1, 3), c(3, 5, NA, NA, 1, NA, NA, 1, 3))
         c.stack <- compress_stack(stack)
}
