% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{compress_pos}
\alias{compress_pos}
\title{Since positions (e.g. "keep" in compress_stack()) are just runs of increasing inds,
there isn't really a need to keep all of them. Rather, just store the start and end
points of the runs.}
\usage{
compress_pos(keep)
}
\arguments{
\item{keep}{vector of non-sparse position inds}
}
\value{
ranges of the incremental runs in keep (can be directly expanded to whatever was in keep)
}
\description{
Example: all.equal(expand_pos(compress_pos(keep)), keep)
}
