% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{compress_stack}
\alias{compress_stack}
\title{Sparse (NA-gapped) matrix compression}
\usage{
compress_stack(stack, sparse.val = NA)
}
\arguments{
\item{stack}{sparse matrix with real values you want to retain}

\item{sparse.val}{(optional) value taking up too much space (i.e. NA default, or 0, or other)}
}
\value{
compressed matrix (list format)
         - pos : linear index within stack
         - vals: non-NA vals within stack
         - m : number of rows in stack
         - n : number of columns in stack
}
\description{
Compresses matrix to its non-NA values and records their positions
to enable perfect reconstruction on-the-fly. Intended for sparse matrices
and parallel operations as well as general memory management.
}
\details{
In the case of feature objects, the profile and position mats can be 
compressed together. 

Note: for spectral matrices, it is rare that all columns have values. As such,
spectral matrices MUST not be transposed; i.e. each column should be
a ppm value and each row is a spectrum. Otherwise ppms will be lost upon 
reconstruction!

Example: c.stack <- compress_stack(rbind(c(NA, NA, 1, 3, 5, 3, NA, 1, NA), c(NA, 1, 3, 5, NA, 3, NA, 1, 3), c(3, 5, NA, NA, 1, NA, NA, 1, 3)))
}
