% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_expand_window.R
\name{corr_expand_window}
\alias{corr_expand_window}
\title{Expand a correlation window around a given peak index}
\usage{
corr_expand_window(xmat, ppm, peakInd, wind)
}
\arguments{
\item{xmat}{A matrix of NMR spectra.}

\item{ppm}{A vector of ppm values corresponding to the columns of xmat.}

\item{peakInd}{The index of the peak around which to expand the window.}

\item{wind}{A vector defining the window around the peak.}
}
\value{
A list with the following elements:
  \item{peak}{The index of the peak used for window expansion.}
  \item{intcorr}{The integral of the correlation within the expanded window.}
  \item{data}{A data frame containing the correlation values, covariance values, and corresponding ppm values within the expanded window.}
  \item{sr}{The STOCSY object for the expanded window.}
  \item{wind}{The original window vector.}
  \item{newWind}{The expanded window vector.}
  \item{corrRbound}{The index of the right boundary of the expanded window in the \code{wind} vector.}
  \item{corrLbound}{The index of the left boundary of the expanded window in the \code{wind} vector.}
}
\description{
Expand a correlation window around a given peak index
}
