% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlation_pocket_pairs.R
\name{correlation_pocket_pairs}
\alias{correlation_pocket_pairs}
\title{correlation_pocket_pairs}
\usage{
correlation_pocket_pairs(
  x,
  ppm,
  ws,
  reg = NULL,
  plotHeatmap = FALSE,
  wdlimit = 0.99,
  noise.width.multiple = 2,
  top.n.peaks = 5,
  rcutoff = 0.5,
  n.cores = 10
)
}
\arguments{
\item{x}{a matrix with numerical values.}

\item{ppm}{a numeric vector of the same length as \code{ncol(x)} specifying the ppm for each column.}

\item{ws}{an integer specifying half the size of the sliding window used to calculate correlations.}

\item{reg}{an optional vector specifying the column indices to consider in \code{x}.}

\item{plotHeatmap}{a logical indicating whether to plot a heatmap of the correlation matrix.}

\item{wdlimit}{a numeric (0,1) specifying the minimum fraction of sliding windows that need to contain a pocket to consider a point noise. Like asking "what 1-the minimum fraction of points you think might be noise?". If at least 5% of the points are noise, choose 0.95}

\item{rcutoff}{a numeric (0,1) specifying the correlation coefficient cutoff to use when extracting significant peaks.}
}
\value{
a list with the following elements:
\item{regions}{a matrix specifying the indices of the peaks in each column of the original matrix, filtered to exclude peaks that are too small or only have unidirectional interactions.}
\item{corr}{a matrix of correlation coefficients between columns of \code{x}.}
\item{cov}{a matrix of covariance values between columns of \code{x}.}
\item{peakMap}{a matrix specifying the indices of the peaks in each column of the correlation matrix.}
\item{noiseDist}{the proportion of pockets containing each window index.}
}
\description{
correlation_pocket_pairs
}
