% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrPocketPairs_al.R
\name{corrPocketPairs_al}
\alias{corrPocketPairs_al}
\title{corrPocketPairs}
\usage{
corrPocketPairs_al(
  x,
  ppm,
  ws,
  reg = NULL,
  plotHeatmap = FALSE,
  wdlimit = 0.99,
  rcutoff = 0.75
)
}
\arguments{
\item{x}{a matrix with numerical values.}

\item{ppm}{a numeric vector of the same length as \code{ncol(x)} specifying the ppm for each column.}

\item{ws}{an integer specifying the size of the sliding window used to calculate correlations.}

\item{reg}{an optional vector specifying the column indices to consider in \code{x}.}

\item{plotHeatmap}{a logical indicating whether to plot a heatmap of the correlation matrix.}

\item{wdlimit}{a numeric specifying the minimum fraction of sliding windows that need to contain a pocket to consider it noise.}

\item{rcutoff}{a numeric specifying the correlation coefficient cutoff to use when extracting significant peaks.}
}
\value{
a list with the following elements:
\item{regions}{a matrix specifying the indices of the peaks in each column of the original matrix, filtered to exclude peaks that are too small or only have unidirectional interactions.}
\item{corr}{a matrix of correlation coefficients between columns of \code{x}.}
\item{cov}{a matrix of covariance values between columns of \code{x}.}
\item{peakMap}{a matrix specifying the indices of the peaks in each column of the correlation matrix.}
\item{noiseDist}{the proportion of pockets containing each window index.}
}
\description{
corrPocketPairs
}
