% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debug_write.R
\name{debug_write}
\alias{debug_write}
\title{debug_write #########################################################################################################################}
\usage{
debug_write(obj, f.name, pars)
}
\arguments{
\item{obj}{some object you want to save if debugging}

\item{f.name}{RDS filename the object should be saved under (no directory info, just name with extension)}

\item{pars}{parameters from the run; contains info on whether of not to save}
}
\description{
If saving all outputs to the pipeline, then save obj under f.name in  directory
Uses pars$debug$enabled to check if debugging. If so, check pars$debug$all.outputs 
to see whether or not to save.
}
\details{
Saving location is pars$dirs$temp /debug_extra.outputs
}
