% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect.baseline.effect.R
\name{detect.baseline.effect}
\alias{detect.baseline.effect}
\title{Detect baseline effect in feature data}
\usage{
detect.baseline.effect(feat, cutoff.corr = 0.99, prom.ratio = 0.3)
}
\arguments{
\item{feat}{A numeric vector containing feature data.}

\item{cutoff.corr}{The maximum allowed correlation between predicted and
observed peak intensities.}

\item{prom.ratio}{The minimum required ratio of peak prominence to feature
intensity range.}
}
\value{
A list containing two logical values. The first indicates whether
  the feature passed the peak prominence filter, and the second indicates
  whether the feature passed the peak fit filter.
}
\description{
Detect baseline effect in feature data
}
