% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_baseline_effect.R
\name{detect_baseline_effect}
\alias{detect_baseline_effect}
\title{Detect baseline effect in feature profile (also indicates monotonic features, which are easy to get with STORM)
Goes through a couple of checks:
-> linearly interpolate across NA gaps
- is the feature a singlet?
- is the peak prominence (max vs. lower bound) of at least one peak > prom.ratio x the entire feature intensity range
- do the valleys and peaks in the feature profile correlate strongly (r > cutoff.corr)? If so, likely a baseline effect.}
\usage{
detect_baseline_effect(feat, cutoff.corr = 0.99, prom.ratio = 0.3)
}
\arguments{
\item{feat}{A numeric vector containing feature data.}

\item{cutoff.corr}{The maximum allowed correlation between predicted and
observed peak intensities.}

\item{prom.ratio}{The minimum required ratio of peak prominence to feature
intensity range.}
}
\value{
A list containing two logical values. The first indicates whether
  the feature passed the peak prominence filter, and the second indicates
  whether the feature passed the peak fit filter.
}
\description{
Detect baseline effect in feature profile (also indicates monotonic features, which are easy to get with STORM)
Goes through a couple of checks:
-> linearly interpolate across NA gaps
- is the feature a singlet?
- is the peak prominence (max vs. lower bound) of at least one peak > prom.ratio x the entire feature intensity range
- do the valleys and peaks in the feature profile correlate strongly (r > cutoff.corr)? If so, likely a baseline effect.
}
