% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distribute.R
\name{distribute}
\alias{distribute}
\title{distribute(workloads, across = ncores)
Load balancing based on some amount of work per task (workloads)
to be distributed as evenly as possible across some number of 
cores (across). Messages to tell what each iteration did. 
Basically, this follows the "Big Rocks" principle (steven covey):
- sort the tasks by amount of work required (descending) 
- iteratively assign each task to the core with the lowest workload}
\usage{
distribute(workloads, across, messages = F)
}
\arguments{
\item{workloads}{vector of workload estimates for each object}

\item{across}{number of cores to distribute objects to}

\item{messages}{whether or not to print the assignment process steps as they occur.}
}
\value{
Plan for distributing objects to balance workload. Core.loads are estimated workload, core.ids are the core assignment for each object.
}
\description{
returns: list with:
          - core.loads: assigned cumulative loads for each core
          - core.ids: core assignments for each task
example: distribute(c(123, 2, 324, 1234, 3212, 32, 3, 500, 23, 234, 3, 3214, 1234), 2)
}
\details{
MTJ 2023
}
