% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{drawScatterScores}
\alias{drawScatterScores}
\title{Makes a plotly scatterplot in the style of browse_evidence().}
\usage{
drawScatterScores(mat, dropRowNames = F, source.name = "scatter")
}
\arguments{
\item{mat}{a numerical matrix of values, can have rownames and colnames.}

\item{dropRowNames}{whether or not to replace the rownames with 1:nrow(mat)}

\item{source.name}{tag for tracking in shiny apps}
}
\value{
interactive plotly heatmap object
}
\description{
Makes a plotly scatterplot in the style of browse_evidence().
}
