% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcorr.R
\name{earlyOut}
\alias{earlyOut}
\title{Return early output for `xcorr` if no comparisons should be computed}
\usage{
earlyOut(dontCalc, overlaps)
}
\arguments{
\item{dontCalc}{A boolean vector indicating which comparisons are not calculated}

\item{overlaps}{A numeric vector of the overlaps between the reference and each
shift of the spectrum}
}
\value{
A list containing NA values for `rvals`, `pvals`, `lags`, and `bestFit`,
  as well as the `overlaps` values and `dontCalc` boolean vector.
}
\description{
This function returns a list of NA values for `rvals`, `pvals`, `lags`, and `bestFit`,
as well as the `overlaps` values and `dontCalc` boolean vector that caused the
function to exit early.
}
