% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcorr_localign.R
\name{errorOut}
\alias{errorOut}
\title{Create an error output object with default values.}
\usage{
errorOut()
}
\value{
A list with the following elements:
\describe{
  \item{best.shifts}{A numeric vector of length 1, set to 0.}
  \item{best.rvals}{A numeric vector of length 1, set to 0.}
  \item{best.pvals}{A numeric vector of length 1, set to 1.}
  \item{shiftedInds}{A NULL object.}
  \item{shiftedSpecs}{A NULL object.}
  \item{ref.aligned}{A NULL object.}
  \item{overlap.sizes}{A numeric vector of length 1, set to 0.}
  \item{g}{A NULL object.}
}
}
\description{
Create an error output object with default values.
}
\examples{
errorOut()

}
