% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/every.nth.R
\name{every.nth}
\alias{every.nth}
\title{Give logical indices to evenly subset "select" elements of "from", an array of something.}
\usage{
every.nth(select, from)
}
\arguments{
\item{select}{target length of the evenly spaced subset of "from"}

\item{from}{length of an array (assumption if length == 1), or the array itself}
}
\value{
logical indices of "from" or the thing that "from" is the length of
}
\description{
Give logical indices to evenly subset "select" elements of "from", an array of something.
}
