% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{expand_pos}
\alias{expand_pos}
\title{Expand ranges to a vector}
\usage{
expand_pos(ranges)
}
\arguments{
\item{keep}{vector of non-sparse position inds}
}
\value{
ranges of the incremental runs in keep

Example: all.equal(expand_pos(compress_pos(keep)), keep)
}
\description{
Expand ranges to a vector
}
