% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_ref.R
\name{expand_ref}
\alias{expand_ref}
\title{expand.ref}
\usage{
expand_ref(ref, ppm)
}
\arguments{
\item{ref}{list of dataset-processed library data for a compound}

\item{ppm}{dataset ppm vector the data were interpolated to}
}
\value{
ref with mapped$data and mapped$ppm expanded, ready for use
}
\description{
Expand library data processed
}
\details{
For a given ref object, expand the compressed data and ppm vector. The ppm
vector will only expand for non-na values; thus, just replace with ppm for 
the dataset (it's just a copy, anyways).
}
