% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expand_window.R
\name{expand_window}
\alias{expand_window}
\title{Expand a window by a certain distance while keeping indices within a given range.}
\usage{
expand_window(window, within, by = NULL)
}
\arguments{
\item{window}{A numeric vector representing the original window to be expanded.}

\item{within}{A numeric vector representing the range within which the expanded window should be contained.}

\item{by}{A numeric scalar representing the amount by which to expand the window. If not specified, the default is to expand the window by the span of its original values.}
}
\value{
A numeric vector representing the expanded window.
}
\description{
Expand a window by a certain distance while keeping indices within a given range.
}
\examples{
# Expand a window with default arguments
expand_window(c(1, 3), 1:5)

# Expand a window with custom expansion distance
expand_window(c(1, 3), 1:5, by = 2)

}
