% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{extract_pos}
\alias{extract_pos}
\title{From a compressed position (in cstack), extract the original position. 
(Utility for hiding annoying sub2ind and ind2sub operations).}
\usage{
extract_pos(cstack)
}
\arguments{
\item{cstack}{compressed stack (or compressed stack.list) containing compressed positions}
}
\value{
positions in original stack, as linear indices
}
\description{
From a compressed position (in cstack), extract the original position. 
(Utility for hiding annoying sub2ind and ind2sub operations).
}
