% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractPeaks_corr.R
\name{extractPeaks_corr}
\alias{extractPeaks_corr}
\title{Extract local maxima and minima from a correlation vector}
\usage{
extractPeaks_corr(corr, mask = NULL, plots = FALSE)
}
\arguments{
\item{corr}{A correlation vector.}

\item{mask}{A boolean mask for excluding points.}

\item{plots}{Whether to plot the resulting peaks and bounds.}
}
\value{
A list containing the peak indices, the bounds of the windows around each peak, and the indices of any masked regions.
}
\description{
This function takes a correlation vector and returns the indices of the local maxima and minima, and the bounds for the corresponding windows around each maximum.
}
\examples{
corr <- rnorm(100)
peaks <- extractPeaks_corr(corr, plots = TRUE)

}
