% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastStack.R
\name{fastStack}
\alias{fastStack}
\title{Fast alternative to stackplot(), geom_area-based plots, or the like.
Uses scattermore or ggplot2, but pre-computes points covered by lower spectra
(i.e. lower row #; foreground) and fills with NA to give the appearance of 
being layered.}
\usage{
fastStack(
  x,
  ppm,
  raster = F,
  vshift = 1,
  pixels = c(512, 512),
  pointsize = 0,
  interpolate = T
)
}
\arguments{
\item{x}{spectral matrix}

\item{ppm}{ppm vector}

\item{raster}{use scattermore? default FALSE gives geom_line based plot. Both return ggplot obj.}

\item{vshift}{how much to shift each consecutive row upwards (% of standard deviation for x)}

\item{pixels}{vector of (rows, cols) - used by scattermore (raster = T)}

\item{pointsize}{~ linewidth - used by scattermore (raster = T). 0 is fastest (1 pixel)}

\item{interpolate}{- used by scattermore (raster = T). Nicer result if (default) TRUE.}
}
\value{
ggplot object; stackplot of spectra
}
\description{
Row 1 is the bottom spectrum in the plot. Note: x is currently  minmax scaled [0,1].
}
