% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat.align.R
\name{feat.align}
\alias{feat.align}
\title{Align features to each other (pairwise) using fft convolution}
\usage{
feat.align(feature.stack, max.hits = 1, counter = T)
}
\arguments{
\item{feature.stack}{matrix of feature profiles (NAs allowed, but will be zero-filled)}

\item{max.hits}{number of top-ranked conv peaks to report}

\item{counter}{whether or not to print message for each feature being aligned}
}
\value{
lag table
}
\description{
Align features to each other (pairwise) using fft convolution
}
