% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feat_align_to.R
\name{feat_align_to}
\alias{feat_align_to}
\title{Align features in "align" to a specified feature
Uses fft convolution. Concatenates features with padding and does a single fft calc.
Note: don't do this for huge feature numbers, or will cost in memory.}
\usage{
feat_align_to(align, to, max.hits = 1, max.lag = Inf)
}
\arguments{
\item{align}{matrix of feature profiles (NAs allowed, but will be zero-filled)}

\item{to}{row index (of matrix) for feature to align to}

\item{max.hits}{number of top-ranked conv peaks to report}

\item{max.lag}{mag lag to consider}
}
\value{
lag table
}
\description{
Align features in "align" to a specified feature
Uses fft convolution. Concatenates features with padding and does a single fft calc.
Note: don't do this for huge feature numbers, or will cost in memory.
}
