% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature.scale.R
\name{feature.scale}
\alias{feature.scale}
\title{Min-max scale the feature profiles (rows) to highest true peak in each.
Uses peaks which were already extracted.}
\usage{
feature.scale(featureStack, peaks)
}
\arguments{
\item{featureStack}{feature profile matrix (each row is a feature)}

\item{peaks}{peaks for each, from feature.extractPeaks}
}
\value{
scaled feature matrix
}
\description{
Min-max scale the feature profiles (rows) to highest true peak in each.
Uses peaks which were already extracted.
}
