% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_est_plot.R
\name{feature_est_plot}
\alias{feature_est_plot}
\title{Plot features onto ref region using selected data from select_evidence_refmet()
Accessory plotting function for browse_evidence(), makes EST-style lines that 
allow user to browse the feature-level evidence for a given reference (instead
of sample-specific). Answers the question: what is the coverage for different
parts of the reference spectrum?}
\usage{
feature_est_plot(reg, metab.evidence, features.c, ppm, plt.pars)
}
\arguments{
\item{reg}{currently plotted ppm region of reference}

\item{metab.evidence}{selected metabolite evidence for given samples}

\item{features.c}{compressed feature obj.}

\item{ppm}{ppm vector}

\item{plt.pars}{plot parameters (main one is plt.pars$pixels)}
}
\value{
scattermore plot with two layers (lines and blocks)
}
\description{
Plot features onto ref region using selected data from select_evidence_refmet()
Accessory plotting function for browse_evidence(), makes EST-style lines that 
allow user to browse the feature-level evidence for a given reference (instead
of sample-specific). Answers the question: what is the coverage for different
parts of the reference spectrum?
}
