% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_extractPeaks.R
\name{feature_extractPeaks}
\alias{feature_extractPeaks}
\title{Extract true peaks from feature using extractPeaks_corr. Formats with peak max
inds easily accessed.}
\usage{
feature_extractPeaks(feature)
}
\arguments{
\item{feature}{feauture profile}
}
\value{
Peak information for true peaks, including
  - pks (all peak info)
  - maxpk highest intensity true peak
  - pkind.max index of maxpk
}
\description{
Extract true peaks from feature using extractPeaks_corr. Formats with peak max
inds easily accessed.
}
