% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/feature_match2ref_slim_ncc.R
\name{feature_match2ref_slim_ncc}
\alias{feature_match2ref_slim_ncc}
\title{Match a feature against a reference using FFT-based convolution and correlation
picks the top max.hits peaks in the convolution, then calculates correlation and
pvalue of the PCC at those points in the ref.
Returns a dataframe with the xcorr information.}
\usage{
feature_match2ref_slim_ncc(
  f.num,
  r.num,
  feat,
  ref,
  feat.ft.c,
  ref.ft,
  pad.size,
  max.hits = 5,
  r.thresh = 0.8,
  p.thresh = 0.01,
  trim = FALSE
)
}
\arguments{
\item{f.num}{numeric: Feature number}

\item{r.num}{numeric: Reference number}

\item{feat}{numeric: Vector of feature intensities}

\item{ref}{numeric: Vector of reference intensities}

\item{feat.ft.c}{numeric: FFT of feature intensities, conjugated}

\item{ref.ft}{numeric: FFT of reference intensities}

\item{pad.size}{numeric: Size of the padding applied during FFT-based convolution}

\item{max.hits}{numeric: Maximum number of candidate lags to consider}

\item{r.thresh}{numeric: Correlation threshold for considering a match}

\item{p.thresh}{numeric: P-value threshold for considering a match}
}
\value{
A data frame with the following columns:
  - feat: Feature number or identifier
  - ref: Reference number or identifier
  - lag: The lag that produces the highest correlation
  - rval: The correlation coefficient
  - pval: The p-value of the correlation
  - pts.matched: Number of data points used in the correlation
  - pts.feat: Number of data points in the feature vector
  - feat.start: The starting index of the feature vector used in the correlation
  - feat.end: The ending index of the feature vector used in the correlation
  - ref.start: The starting index of the reference vector used in the correlation
  - ref.end: The ending index of the reference vector used in the correlation
}
\description{
Match a feature against a reference using FFT-based convolution and correlation
picks the top max.hits peaks in the convolution, then calculates correlation and
pvalue of the PCC at those points in the ref.
Returns a dataframe with the xcorr information.
}
