% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fillbetween.R
\name{fillbetween}
\alias{fillbetween}
\title{fillbetween}
\usage{
fillbetween(inds)
}
\arguments{
\item{inds}{a numeric vector of length 2 indicating the starting and ending indices}
}
\value{
a numeric vector representing the sequence of indices between the starting and ending indices
}
\description{
Calculate a sequence of values between two numeric indices, with optional reversal.
}
\examples{
fillbetween(c(1, 5)) # returns 1 2 3 4 5
fillbetween(c(5, 1)) # returns 5 4 3 2 1
fillbetween(c(1, NA)) # throws an error
}
