% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.matches.R
\name{filter.matches}
\alias{filter.matches}
\title{Match Filtering}
\usage{
\method{filter}{matches}(pars)
}
\arguments{
\item{pars}{A list of input parameters.}
}
\value{
A list of filtered and processed matched peak information, including back-fits to the original spectra.
}
\description{
Filter and process matched peaks based on user-specified criteria, such as singlet removal and ppm distance filtering.
}
