% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.matches_singlets.R
\name{filter.matches_singlets}
\alias{filter.matches_singlets}
\title{Filter matches with singlets}
\usage{
\method{filter}{matches_singlets}(
  match.info,
  fits.feature,
  peak.qualities,
  pq.featureNumbers,
  res.area.thresh
)
}
\arguments{
\item{match.info}{The match information data.frame.}

\item{fits.feature}{A list of fitted features.}

\item{peak.qualities}{A list of peak qualities.}

\item{pq.featureNumbers}{A vector of feature numbers.}

\item{res.area.thresh}{The threshold for residual area.}
}
\value{
A list containing the filtered match information and the filtered fitted features.
}
\description{
This function removes matches that have only one peak in their respective feature, reference or feature-not-never-fit regions.
}
