% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_matches.R
\name{filter_matches}
\alias{filter_matches}
\title{Match Filtering}
\usage{
filter_matches(pars)
}
\arguments{
\item{pars}{A list of input parameters.}
}
\value{
A list of filtered and processed matched peak information, including back-fits to the original spectra.
}
\description{
Filter and process matched peaks based on user-specified criteria, such as singlet removal and ppm distance filtering.
Also calculate backfits of ref-feats (reference subsignatures fished out by feature matches to reference spectra) to each individual dataset spectrum. 
Note: backfits are now calculated directly from ref features and position in 
ss.spec is optimized using updated batman-style fitting. This strategy has proven
to be much more accurate and produces meaningful scores.
Note: BFF (backfit feasibility) scores are no longer used. Instead, updated batman fits
yield a fraction.spec.accounted score which quantifies the fraction of relevant spectral
signal accounted for by the ref feature. positive overshoot is penalized by 
subtracting from the score. This is much more simple and intuitive.
}
