% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_matches_singlets.R
\name{filter_matches_singlets}
\alias{filter_matches_singlets}
\title{Filter matches with singlets}
\usage{
filter_matches_singlets(
  match.info,
  f.cstack,
  r.cstack,
  res.area.threshold,
  ncores
)
}
\arguments{
\item{match.info}{The match information data.frame obtained within filter.matches.}

\item{feature.stack}{e.g. feature$stack}

\item{ref.mat}{spectral matrix for reference compounds}

\item{peak.qualities}{A list of peak quality vectors (~ feature points' relevance in the reference database)}

\item{pq.featureNumbers}{A vector of feature numbers to index the peak.qualities list. This is necessary when only a subset of features are matched and feature fits are filtered.}

\item{res.area.thresh}{The minimum reference spectrum resonance area that must be accounted for by the fit feature in order to consider it matched.}
}
\value{
A list containing the filtered match information and the filtered fitted features.
}
\description{
This function removes matches that have only one peak in their respective feature, reference or feature-not-never-fit regions.
}
