% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filterFeatures.R
\name{filterFeatures}
\alias{filterFeatures}
\title{Filter features based on specified criteria
- null features
- features outside of ppm range
- features with no runs >= min.runlength
- features derived from < min.subset spectra
- features with strong baseline effect}
\usage{
filterFeatures(
  feature,
  ppm,
  ppm.range,
  min.runlength = 3,
  min.subset = 5,
  prom.ratio = 0.3,
  give = "filter"
)
}
\arguments{
\item{feature}{A feature object to be filtered}

\item{ppm}{A vector of ppm values for each point in the spectra}

\item{ppm.range}{A vector specifying the range of ppm values to include}

\item{min.runlength}{The minimum length of a run of consecutive data points to include}

\item{min.subset}{The minimum number of spectra a feature must be present in to be included}

\item{prom.ratio}{The maximum ratio of peak prominence to intensity range for a feature to be considered monotonic}

\item{give}{A character string indicating whether to return the filtered feature object or the filter as a logical vector}
}
\value{
If \code{give = "features"}, a feature object with the filtered features. If \code{give = "filter"}, a logical vector indicating which features passed the filter.
}
\description{
Filter features based on specified criteria
- null features
- features outside of ppm range
- features with no runs >= min.runlength
- features derived from < min.subset spectra
- features with strong baseline effect
}
