% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitFeature.R
\name{fit.minmax}
\alias{fit.minmax}
\title{Fit a feature to a spectrum using minmax scaling}
\usage{
fit.minmax(feature, spectrum)
}
\arguments{
\item{feature}{a numeric vector representing the feature to be fit}

\item{spectrum}{a numeric vector representing the spectrum to which the feature will be fit}
}
\value{
a list with the following components:
\describe{
  \item{feature.fit}{a numeric vector representing the feature fit}
  \item{ratio}{a ratio of the original feature intensity to the intensity of the feature after minmax scaling}
}
}
\description{
Fit a feature to a spectrum using minmax scaling
}
