% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_methods.R
\name{fit_leastSquares}
\alias{fit_leastSquares}
\title{Fit least squares model between two signals and calculate statistics}
\usage{
fit_leastSquares(v1 = NA, v2 = NA, ppm = NULL, plots = FALSE, scale.v2 = TRUE)
}
\arguments{
\item{v1}{a numeric vector representing one of the signals}

\item{v2}{a numeric vector representing the other signal}

\item{ppm}{a numeric vector representing the ppm values for the signals}

\item{plots}{a logical value indicating whether to produce a plot (default is \code{FALSE})}
}
\value{
A list with the following elements:
}
\description{
Fit least squares model between two signals and calculate statistics
}
