% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwhm.R
\name{fwhm}
\alias{fwhm}
\title{Full Width at Half Maximum (FWHM) calculation}
\usage{
fwhm(isolatedpk)
}
\arguments{
\item{isolatedpk}{a numeric vector representing an isolated peak}
}
\value{
a numeric value representing the FWHM of the peak
}
\description{
Given an isolated peak, calculates its Full Width at Half Maximum (FWHM),
i.e., the width of the peak at half of its maximum height.
}
