% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_plot_specfits.R
\name{grid_plot_specfits}
\alias{grid_plot_specfits}
\title{Plot a grid of ref features fit to spectra.}
\usage{
grid_plot_specfits(
  sfs,
  feature,
  xmat,
  refmat,
  plotLoc = "./",
  filename = "grid_specfits",
  titles = "number"
)
}
\arguments{
\item{sfs}{long format spec feature information (includes backfit information)}

\item{feature}{features object for study}

\item{xmat}{spectral matrix for study}

\item{refmat}{reference spectral matrix for study (spectra on rows)}

\item{plotLoc}{plot location, ends with "/"}

\item{filename}{filename, excluding ".pdf"}

\item{titles}{options for data to display on title of each subplot: 
- 'number' (just the plot number, for tracking)
- 'score' (just the score)
- 'number_score' (both)}
}
\value{
grid plot written to pdf specified
}
\description{
Plot a grid of ref features fit to spectra.
}
