% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind2subR.R
\name{ind2subR}
\alias{ind2subR}
\title{Convert linear indices to row and column subscripts}
\usage{
ind2subR(inds, m)
}
\arguments{
\item{inds}{A numeric vector of linear indices}

\item{m}{A positive integer representing the number of columns in the corresponding matrix}
}
\value{
A list with two elements: \code{rows}, a vector of row subscripts corresponding to the linear indices, and \code{cols}, a vector of column subscripts corresponding to the linear indices
}
\description{
This function takes a numeric vector of linear indices \code{inds} and a positive integer \code{m}, representing the number of columns in a matrix, and returns a list with two elements: \code{rows}, a vector of row subscripts corresponding to the linear indices, and \code{cols}, a vector of column subscripts corresponding to the linear indices.
}
\examples{
inds <- 1:9
m <- 3
ind2subR(inds, m)
# Returns: list(rows = c(1, 2, 3, 1, 2, 3, 1, 2, 3), cols = c(1, 1, 1, 2, 2, 2, 3, 3, 3))
}
