% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_indexing.R
\name{index_studies}
\alias{index_studies}
\title{Indexes unzipped studies found in a subdirectory}
\usage{
index_studies(data.dir, exclude = NULL)
}
\arguments{
\item{data.dir}{directory containing zip files of your run results}

\item{exclude}{(optional) list of run names that are problematic (i.e. cannot be unzipped correctly for some reason, or which you don't want to include in the index)}
}
\value{
data frame containing the compiled run summary data for the studies in the directory. Also written to ./run.summary.csv.
}
\description{
Sister function to unzip_studies()
}
\details{
After running unzip_studies(), it may be useful to get an overview of all the results in the directory.
This function compiles all run summaries into a single data frame, and saves the table in a csv file in the directory. 

Examples:
 run.idx <- index_studies(data.dir, exclude = c('1697751863','1697759923','1713439191',''))
}
