% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_true_peak.R
\name{is_true_peak}
\alias{is_true_peak}
\title{Label True Peak points in a vector using extractPeaks_corr (wrapper)}
\usage{
is_true_peak(v)
}
\arguments{
\item{v}{a vector of intensities}
}
\value{
a logical vector indicating the true peaks and tail regions
}
\description{
Label True Peak points in a vector using extractPeaks_corr (wrapper)
}
\examples{
v <- rnorm(100)
is.true.peak(v)

}
