% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lag_features.R
\name{lag_features}
\alias{lag_features}
\title{From a feature stack and lag table, align features to a given feature. Works
on pairs as well. Returns the matrix containing aligned profiles (e.g., for plotting).}
\usage{
lag_features(featureStack, lag.table, to = NULL)
}
\arguments{
\item{featureStack}{matrix containing feature profiles}

\item{lag.table}{lag information table from feat_align or similar}

\item{to}{feature number to align to (must be in f1 or f2)}
}
\value{
matrix with aligned feature profiles.
}
\description{
lag.table example:
  f1 f2 lag.in.f2 pos.big          val
  1  1  2         0    9172 3.705236e+18
  2  2  1         0    3057 3.705236e+18
(last 2 columns unused here, but are in the output from feat_align())
}
\details{
lag.table f1 and f2 values need to index rows of featureStack.
lag.table is the key to extract rows from featureStack and align them.
}
