% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/localMaxima.R
\name{localMaxima}
\alias{localMaxima}
\title{Find the indices of the local maxima of a numeric vector}
\usage{
localMaxima(v)
}
\arguments{
\item{v}{A numeric vector}
}
\value{
A numeric vector containing the indices of the local maxima of \code{v}
}
\description{
Given a numeric vector, this function returns the indices of the local maxima.
A local maximum is defined as a data point that is greater than its adjacent
points. The function adds a minimum value to each end of the vector to ensure that the
endpoints can also be identified as local maxima, if appropriate.
}
\examples{
localMaxima(c(1, 2, 3, 2, 1))

}
