% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.features2refs.par.R
\name{match.features2refs.par}
\alias{match.features2refs.par}
\title{Match STORM features to GISSMO database using parallel computing}
\usage{
match.features2refs.par(pars)
}
\arguments{
\item{pars}{a list of parameters for matching, read in from a YAML file}
}
\value{
nothing; results are saved as an RDS file
}
\description{
This function matches STORM features to a reference database using cross-correlation,
least-squares fitting, and peak quality weighting. The matching process is parallelized
for speed, and results are saved as an RDS file. The function reads in parameters from
a YAML file and data from RDS files generated by a previous function call.
}
