% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.features2refs.par.explicit.setup.R
\name{match.features2refs.par.setup}
\alias{match.features2refs.par.setup}
\title{match.features2refs.par.explicit.setup function}
\usage{
match.features2refs.par.setup(pars)
}
\arguments{
\item{pars}{a list of input parameters}
}
\value{
None
}
\description{
Description:
Description
}
