% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_features2refs_par_explicit.R
\name{match_features2refs_par_explicit}
\alias{match_features2refs_par_explicit}
\title{match.features2refs.par.explicit}
\usage{
match_features2refs_par_explicit(pars)
}
\arguments{
\item{pars}{A list of parameters containing directories, parallel settings, and matching settings.}
}
\value{
matches.RDS file containing match.info, fits, and peak quality data.
}
\description{
This function matches features to reference spectra using a parallelized loop. 
Feature chunking strategy. 
Do you match features to refs, or refs to feature? If using batman-style fit, 
since refs are more pure, perhaps the latter makes more sense. In other words,
fit the cleaner signals to the dirtier ones if using batman. For less biased,
use least squares.
}
\author{
MTJ
}
