% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_features2refs_par_setup.R
\name{match_features2refs_par_setup}
\alias{match_features2refs_par_setup}
\title{match.features2refs.par.explicit.setup function}
\usage{
match_features2refs_par_setup(pars)
}
\arguments{
\item{pars}{a list of input parameters}
}
\value{
writes necessary files for matching script
}
\description{
Chunks feature data and precomputes ref ffts for match.features2refs.par.explicit.
}
