% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchCluster2.R
\name{matchCluster2}
\alias{matchCluster2}
\title{matchCluster2}
\usage{
matchCluster2(t, r, tol = 0.02, method = "itmin")
}
\arguments{
\item{t}{target peaks.}

\item{r}{reference peaks.}

\item{tol}{tolerance in ppm.}

\item{method}{matching methodology as a string, defaults to 'itmin'.
Can also be substructure based 'substruct'. optimal 'opt' not implemented.}
}
\value{
List of matches.
}
\description{
description
}
