% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchToMultiRef.R
\name{matchToMultiRef}
\alias{matchToMultiRef}
\title{matchToMultiRef}
\usage{
matchToMultiRef(
  references,
  metadata,
  target,
  driver_ppm,
  matchMethod = "",
  tol = 0.02,
  Itol = 20,
  intensity = FALSE
)
}
\arguments{
\item{references}{reference spectra as a db file.}

\item{metadata}{metadata that is included in the final merged dataframe.}

\item{target}{target driver peak.}

\item{driver_ppm}{parts per million of driver peak.}

\item{tol}{tolerance in ppm.}

\item{Itol}{tbc.}

\item{intensity}{flag indicating whether to figure intensity into matching.}

\item{match_method}{matching methodology as a string.}
}
\value{
Dataframe of results ranked by score.
}
\description{
Matches peaks from one pseudo-spectrum corresponding to one driver peak
to all database peaks.
Returns a data frame with the results ranked by score.
}
