% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mem_snapshot.R
\name{mem_snapshot}
\alias{mem_snapshot}
\title{Write a file with current R memory usage}
\usage{
mem_snapshot(fname)
}
\arguments{
\item{fname}{filename to write to (usually timestamped)}
}
\value{
writes a file with current R memory usage
}
\description{
Write a file with current R memory usage
}
