% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_specFit.R
\name{opt_specFit}
\alias{opt_specFit}
\title{Given a specFit row (match info plus ss.spec information), expand the matched
spec region to span-1 on each side of the specfit. For each internal lag, do 
batman-style fitting and assess based on climbing to the first local maximum 
spec signal accounted for (starting from current fit position). Return the 
updated specFit row.}
\usage{
opt_specFit(sf, feature, xmat, refmat)
}
\arguments{
\item{sf}{spec feature information (includes backfit information)}

\item{feature}{features object for study}

\item{xmat}{spectral matrix for study}

\item{refmat}{reference spectral matrix for study (spectra on rows)}
}
\value{
updated specFit row (sf), with optimized fit coefficients and spec start/end
}
\description{
Note: this is meant to be run during backfitting. Using after that will not 
update some fields calculated there, e.g.:
}
\details{
"bffs.res" "bffs.tot" "rmse" "rmse.biased" "pct.ref"
}
