% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/padmat.R
\name{padmat}
\alias{padmat}
\title{Pad matrix with rows and columns of specified value}
\usage{
padmat(x, use = NA, row.by = 0, col.by = 1)
}
\arguments{
\item{x}{A matrix to pad}

\item{use}{The value to fill new rows and columns with (default is NA)}

\item{row.by}{The number of rows to add to each side of the matrix}

\item{col.by}{The number of columns to add to each side of the matrix}
}
\value{
A padded matrix
}
\description{
This function takes a matrix and pads it with rows and columns of a specified
value. The number of rows and columns to add can also be specified. The new 
rows and columns will have the same value.
}
\examples{
m <- matrix(1:4, ncol = 2)
padmat(m, use = 0, row.by = 1, col.by = 1)

m2 <- matrix(1:6, ncol = 2)
padmat(m2, row.by = 2, col.by = 3)

m3 <- matrix(1:9, ncol = 3)
padmat(m3, use = -1, row.by = 1, col.by = 2)

}
\seealso{
\code{\link{sub2indR}}
}
\keyword{manipulations}
\keyword{matrix}
