% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Pipeline function to execute the entire ImperialNMRTool workflow}
\usage{
pipeline(params_loc, params_obj)
}
\arguments{
\item{params_loc}{Path to a YAML file containing user-specified parameters}

\item{params_obj}{List of parameters}
}
\description{
Feature shape based annotation pipeline
 /
 To run from here, you need the following files in ./data:
 params.yaml         # parameter file
 spectral.matrix.RDS # spectral matrix
 data.list.RDS       # library files
 lib.info.RDS        # index for library files
}
\details{
Running the setup script will accomplish this, as well as ensuring that the correct
 library is being used.
 This pipeline accepts a spectral matrix (preferably with minimal processing; e.g.
 no strong normalization, no scaling, no alignment) in the form of and RDS file
 with a matrix (row 1 = ppm vector, each additional row is a 1D NMR sample).

 FSE: the spectral matrix is decomposed into compound features using feature
 shape extraction. First, a local STOCSY is performed at every point along the
 spectrum (within a sliding window of ~ 100 points; enough to capture multiple
 resonances within any multiplet). For each of these STOCSYs, the central peak
 in the correlation profile (correlation pocket; corrpocket) typically captures
 a resonance, as the correlation is 1 by definition at the STOCSY'd point, and
 typically falls off as you approach the boundaries of the resonance. This is
 taken, with the next highest correlation peak within the window, to form a
 rough statistical description of two resonances which have an correlation in
 intensity across samples, albeit separated by chemical shift. We term this a
 'protofeature'. Importantly, each point and its associated window will capture
 the dominant protofeature most associated with that point. This changes for
 adjacent points, and many protofeatures will be duplicated multiple times. A
 protofeature should be considered as a rough hypothesis about a statistical
 association between two resonances, which happen to be sufficiently aligned
 so that they produce a coherent signal.
 If the windows are all aligned, we can plot the % of central corrpeaks containing
 each window point. From this distribution, it is clear that nearly all
 protofeatures, including those from noise peaks and real peaks, include the
 most central window points. As such, these cannot reliably be used to identify
 noise. However, the correlation peak about noise tends to be much smaller, and
 characteristically so. As such, a noisewidth can be estimated from this
 distribution. This is the origin of the noise.percentile cutoff, which is applied
 like so: "given a noise.percentile = 0.99, consider only those protofeatures
 for which both peaks have a width > the smallest 1% of peaks". Reducing this
 number therefore gives a more selective cut. Protofeatures are also filtered so
 that the

 STORM: Joram Posma's STORM has been adapted and optimized to accept these
 protofeatures in the following ways:
 - first, since many of the protofeatures are noise,

# MTJ 2023
}
