% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pk.maxs.R
\name{pk.maxs}
\alias{pk.maxs}
\title{Get indices of local maxima in a vector}
\usage{
pk.maxs(v, mask)
}
\arguments{
\item{v}{A numeric vector}

\item{mask}{A logical vector of the same length as \code{v}, indicating which values of \code{v} should be masked (i.e., excluded from consideration as local maxima)}
}
\value{
A numeric vector containing the indices of the local maxima in \code{v} that are not masked by \code{mask}
}
\description{
This function takes a numeric vector \code{v} and a logical mask \code{mask}, and returns the indices of the local maxima in \code{v} that are not masked by \code{mask}.
}
\examples{
v <- c(1, 2, 3, 2, 1)
mask <- c(FALSE, FALSE, TRUE, FALSE, FALSE)
pk.maxs(v, mask)
# Returns: 3
}
