% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fit.R
\name{plot.fit}
\alias{plot.fit}
\title{Plot the spectral and feature fits}
\usage{
\method{plot}{fit}(fit, type = "simple", ppm = NULL, color = NULL)
}
\arguments{
\item{fit}{An object of class "fit"}

\item{type}{A character string specifying the type of plot. Possible values are "simple", "auc", and "color.line".}

\item{ppm}{A numeric vector containing the ppm values.}

\item{color}{A numeric vector containing the color values.}
}
\value{
A ggplot object
}
\description{
Plot the spectral and feature fits
}
\examples{
plot.fit(fit, type = "simple", ppm = ppm, color = color)

}
