% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_addRef.R
\name{plot_addRef}
\alias{plot_addRef}
\title{Add reference region to an existing ggplot object}
\usage{
plot_addRef(g, specRegion, ref, ppm, scaling = TRUE, type = "bold line")
}
\arguments{
\item{g}{existing ggplot object (e.g. from simplePlot)}

\item{specRegion}{x[,ref.idx]}

\item{ref}{entire ref object - this function uses information from window and signature}

\item{ppm}{ppm[,ref$wind$inds]}

\item{scaling}{logical, whether to use new scaling}

\item{type}{character vector specifying the type of plot to make ("bold line", "line", "shaded area")}
}
\value{
ggplot object with reference region added
}
\description{
Add reference region to an existing ggplot object
}
