% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_addSTOCSYLine.R
\name{plot_addSTOCSYLine}
\alias{plot_addSTOCSYLine}
\title{Plot STOCSY line over the current plot}
\usage{
plot_addSTOCSYLine(
  g,
  specRegion,
  ppm,
  corr = NULL,
  covar = NULL,
  driver = NULL,
  bounds = NULL
)
}
\arguments{
\item{g}{plot to add STOCSY line to}

\item{specRegion}{region of the NMR spectrum to plot}

\item{ppm}{vector of chemical shift values}

\item{corr}{matrix of correlations between reference and query spectra}

\item{covar}{matrix of covariances between reference and query spectra}

\item{driver}{chemical shift value of the driver peak, if given}

\item{bounds}{chemical shift values of the boundaries of the region of interest, if given}
}
\value{
plot with STOCSY line added
}
\description{
Plot STOCSY line over the current plot
}
