% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation_pocket.R
\name{plot_correlation_pocket}
\alias{plot_correlation_pocket}
\title{Plot correlation window for a given pocket}
\usage{
plot_correlation_pocket(vals)
}
\arguments{
\item{vals}{a list containing values for the correlation window including sr, wind, corrRbound, and corrLbound}
}
\value{
a ggplot object displaying the correlation window
}
\description{
Plot correlation window for a given pocket
}
