% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation_pockets_grid.R
\name{plot_correlation_pockets_grid}
\alias{plot_correlation_pockets_grid}
\title{Plot correlation pockets in a grid}
\usage{
plot_correlation_pockets_grid(allvals, plotLoc)
}
\arguments{
\item{allvals}{a list of correlation pocket data (output from \code{\link{find_correlation_pockets}})}

\item{plotLoc}{character string specifying the directory to save the resulting PDF file}
}
\value{
None
}
\description{
Wrapper function for \code{\link{plot_correlation_pocket}} to generate a grid of plots
from a list of correlation pocket data. Prints the resulting grid to a PDF file
located in the specified directory.
}
