% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrPocketPairs.R
\name{plot_corrPocketPairs}
\alias{plot_corrPocketPairs}
\title{Plot Correlation Heatmaps of Pocket Pairs}
\usage{
plot_corrPocketPairs(
  pocketPairs,
  region = NULL,
  vshift = 5,
  xvect = NULL,
  show = TRUE
)
}
\arguments{
\item{pocketPairs}{A list of pocket pairs with elements named "peakMap", "corr", "cov", and "regions".}

\item{region}{A vector of integers specifying the columns of the peakMap to be included in the plot.}

\item{vshift}{The vertical shift to apply to the distributions for stacking.}

\item{xvect}{The x values at which to plot the distributions.}

\item{show}{Logical. Whether or not to display the plot.}
}
\value{
The correlation heatmap plot.
}
\description{
This function plots correlation heatmaps for a subset of pocket pairs from a list of pocket pairs.
}
\examples{
# Load data
data(pocketPairsList)

# Plot correlation heatmaps
plot_corrPocketPairs(pocketPairsList[[1]], region = 2:5)


}
