% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrPocketPairs_grid.R
\name{plot_corrPocketPairs_grid}
\alias{plot_corrPocketPairs_grid}
\title{Plot correlation pairs of pockets in a grid}
\usage{
plot_corrPocketPairs_grid(xmat, ppm, pocketPairs, res = 4000, plotLoc = "./")
}
\arguments{
\item{xmat}{A matrix of dimensions (n, m), where each row is a pocket}

\item{ppm}{A vector of ppm values}

\item{pocketPairs}{A list of 2-tuples, each of which indicates which pockets to compare}

\item{res}{Resolution of the plot in ppm}

\item{plotLoc}{A string specifying the directory to save the plot in}
}
\value{
A PDF file of correlation plots between the specified pocket pairs arranged in a grid
}
\description{
Plot correlation pairs of pockets in a grid
}
