% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotly_functions.R
\name{plot_spec}
\alias{plot_spec}
\title{Plot spectrum in plotly}
\usage{
plot_spec(spec, ppm, aucs = NULL, title = "", source.name = "refspec")
}
\arguments{
\item{spec}{spectrum (vector or 1d matrix)}

\item{ppm}{ppm vector for spectrum}

\item{aucs}{area under the curves (not used currently)}

\item{title}{plot title (i.e. compound name)}

\item{source.name}{source event name (for shiny interactive input passing)}
}
\value{
A plotly figure.
}
\description{
plot reference spectrum (or any spectrum) using plotly, with event handler for region selection.
}
