% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stocsy_corrWindow.R
\name{plot_stocsy_corrWindow}
\alias{plot_stocsy_corrWindow}
\title{Plot a correlation window with local minima}
\usage{
plot_stocsy_corrWindow(sr, lbound, rbound, showplot = TRUE, restrictTo = NULL)
}
\arguments{
\item{sr}{An object of class Spectrum containing correlation information}

\item{lbound}{An integer specifying the left boundary of the correlation window to be plotted}

\item{rbound}{An integer specifying the right boundary of the correlation window to be plotted}

\item{showplot}{A logical indicating whether the plot should be displayed on the screen (default = TRUE)}

\item{restrictTo}{A numeric vector of length two specifying the limits of the x-axis in the plot}
}
\value{
A ggplot2 object of the correlation window with local minima highlighted
}
\description{
Plot a correlation window with local minima
}
