% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_storm_refRegions.R
\name{plot_storm_refRegions}
\alias{plot_storm_refRegions}
\title{Plot the STORM results}
\usage{
plot_storm_refRegions(
  xmat = NULL,
  ppm,
  s,
  bgplot = "overlayed",
  vshift = 1,
  hshift = 0,
  calcStocsy = TRUE,
  n_xticks = 4,
  plot.title = "",
  xlim = NULL
)
}
\arguments{
\item{xmat}{numeric matrix of spectral data}

\item{ppm}{numeric vector of length equal to ncol(xmat)}

\item{s}{STORM results from storm_play or similar function}

\item{bgplot}{a character vector indicating how the plot is rendered:
"overlayed" or "stack". Default is "overlayed".}

\item{vshift}{numeric scalar indicating the vertical shift between spectra when bgplot is "stack". Default is 1.}

\item{hshift}{numeric scalar indicating the horizontal shift between spectra when bgplot is "stack". Default is 0.05.}

\item{calcStocsy}{a logical indicating whether to calculate STOCSY plot. Default is TRUE.}

\item{n_xticks}{an integer indicating the number of ticks on the x-axis.
Default is 4.}
}
\value{
a ggplot object with the plot of the spectral data, the reference region, and the STOCSY plot (if applicable).
}
\description{
Plot the storm results (s) from storm_play, etc.
Plot light gray ref region for optimal subset.
use in lapply().
Pass in s:
  nothing outside
}
\examples{
plot_storm_refRegions(xmat = xmat, ppm = ppm, s = stormResults[[1]])

}
\seealso{
\code{\link{plot_stormRefRegions_grid}}, \code{\link{plot_stormRefRegions_summarize}}
}
