% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_storm_refRegions_onlyFeat.R
\name{plot_storm_refRegions_onlyFeat}
\alias{plot_storm_refRegions_onlyFeat}
\title{Plot storm Results for reference regions in grid format - only plot feature profile (not subsets) to reduce weight.}
\usage{
plot_storm_refRegions_onlyFeat(ppm, s, n_xticks = 4)
}
\arguments{
\item{ppm}{a numeric vector of ppm values corresponding to the NMR spectra.}

\item{s}{storm output object (element of list containing the storm analysis results for each reference region, from fse.R)}

\item{n_xticks}{an integer indicating the number of x-axis ticks to use in the plots.}
}
\value{
ggplot object, e.g. to place in grid
}
\description{
Wrapper for \code{\link{plot_storm_refRegions}} to generate plots for multiple
reference regions in a grid format and save the result to a pdf file.
}
