% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_storm_refRegions_grid.R
\name{plot_stormRefRegions_grid}
\alias{plot_stormRefRegions_grid}
\title{Plot storm Results for reference regions in grid format}
\usage{
plot_stormRefRegions_grid(
  xmat,
  ppm,
  stormResults,
  plotLoc,
  filename = "storm_results_refPlots_grid.pdf",
  calcStocsy = TRUE,
  n_xticks = 4
)
}
\arguments{
\item{xmat}{a numeric matrix of NMR spectra with ppm values in the first column.}

\item{ppm}{a numeric vector of ppm values corresponding to the NMR spectra.}

\item{stormResults}{a list containing the storm analysis results for each reference region.}

\item{plotLoc}{a character string indicating the directory to save the output file.}

\item{filename}{a character string of the output filename. Defaults to "storm_results_refPlots_grid.pdf".}

\item{calcStocsy}{a logical indicating whether to calculate and include STOCSY scores in the plots.}

\item{n_xticks}{an integer indicating the number of x-axis ticks to use in the plots.}
}
\value{
a grid of plots of the storm analysis results for each reference region, saved to a pdf file.
}
\description{
Wrapper for \code{\link{plot_storm_refRegions}} to generate plots for multiple
reference regions in a grid format and save the result to a pdf file.
}
