% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_scores_df.R
\name{plot_umap.scores.df}
\alias{plot_umap.scores.df}
\title{Plot UMAP scores}
\usage{
plot_umap.scores.df(df, clusters = NULL)
}
\arguments{
\item{df}{A data.frame with columns "x", "y", and "cluster".}

\item{clusters}{An optional vector of cluster labels for each point in the layout.}
}
\value{
A plotly figure.
}
\description{
Plots UMAP scores (layout) and labels clusters (if given).
}
