% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umap_scores.R
\name{plot_umap.scores}
\alias{plot_umap.scores}
\title{Plot UMAP Scores}
\usage{
plot_umap.scores(umap.obj, clusters = NULL)
}
\arguments{
\item{umap.obj}{A UMAP object generated using the umap package.}

\item{clusters}{A vector of cluster labels for each row of umap.obj$layout.}
}
\value{
A plotly plot of the UMAP scores, with optional cluster labels.
}
\description{
This function plots the UMAP scores (layout) and optionally labels the clusters
(if given).
}
