% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points.covered.R
\name{points.covered}
\alias{points.covered}
\title{Gives logical mask for points covered up by points rows whose index is lower. 
By replacing the TRUE points with NA, one can give the effect of plotted spectra
being layered over top of each other (such that the AUCs of lower row inds 
cover up points in the higher rows). In other words, make an overlay plot with
no criss-crossed lines without introducing unnecessary figure size and slowdown.}
\usage{
\method{points}{covered}(mat)
}
\arguments{
\item{mat}{xmat subset (row and col) that will be plotted}

\item{feats}{features lined up with cols of mat. Corresponding row in mat (for each feature) is given in ss.rows.}

\item{ss.rows}{row in mat for each feature (row in feats) - often duplicates in this vector}
}
\value{
filter for the features matrix
}
\description{
Intended that row 1 is the bottom spectrum in the plot.
}
\author{
MTJ
}
