% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/points_covered_feats.R
\name{points_covered_feats}
\alias{points_covered_feats}
\title{Indicate which points are covered up by a lower spectrum (lower row number)
Use to filter add NAs to a v-and-h-shifted spectral matrix, avoids line crossings 
in stackplots (without explicit layering of area plots). 
* This version also accepts a features matrix (of same size as mat), and ss.rows
to prevent features showing up where spectra lower in the stack are covering them.}
\usage{
points_covered_feats(mat, feats, ss.rows)
}
\arguments{
\item{mat}{xmat subset (row and col) that will be plotted}

\item{feats}{features lined up with cols of mat. Corresponding row in mat (for each feature) is given in ss.rows.}

\item{ss.rows}{row in mat for each feature (row in feats) - often duplicates in this vector}
}
\value{
filter for the features matrix
}
\description{
Intended that row 1 is the bottom spectrum in the plot.
}
