% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ppTargetSpec.R
\name{ppTargetSpec}
\alias{ppTargetSpec}
\title{ppTargetSpecR}
\usage{
ppTargetSpec(target, normalise = FALSE)
}
\arguments{
\item{target}{target spectra to peak pick.}

\item{normalise}{flag indicating normalisation of output, default FALSE.}
}
\value{
list of chemical shifts and intensities.
}
\description{
Function to peak-pick spectra decomposed from mixture spectra.
Requires function detectSpecPeaks from speaq package returns a
list of chemical shifts and intensities which can be optionally
normalised to the most intense peak for improved comparison
against HMDB spectra.
Goncalo Graca, 8 January 2021, g.gomes-da-graca-at-imperial.ac.uk
}
