% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_html_heatmap_from_scoresfile.R
\name{print_html_heatmap_from_scoresfile}
\alias{print_html_heatmap_from_scoresfile}
\title{print_html_heatmap_from_scoresfile}
\usage{
print_html_heatmap_from_scoresfile(tmpdir)
}
\arguments{
\item{pars}{a list containing necessary parameters for the function}
}
\value{
RDS file containing match scores between library reference spectra and the best subset spectrum score for each.
}
\description{
Accessory function to generate and print a heatmap in the browse_evidence style
(using drawHeatmap()) and print it to a self-contained html file.
Note: requires pandoc to be installed if running on Linux systems. If it is not
installed, the file will be written in a different format (YAML metadata e.g. 
for Jekyll??) that browsers won't interpret. If this is the case, you can run
check_heatmap_html_output() on the zip file for the results
}
