% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/printTime.R
\name{printTime}
\alias{printTime}
\title{Accessory function to print timestamp to message throughout pipeline}
\usage{
printTime()
}
\value{
timestamp (message)
}
\description{
Accessory function to print timestamp to message throughout pipeline
}
