% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project_features.stackplot.R
\name{project_features.stackplot}
\alias{project_features.stackplot}
\title{Generate a stacked area plot of the spectra for each feature in \code{xmat}, overlaid with the corresponding best-fit features}
\usage{
project_features.stackplot(
  xmat,
  ppm,
  label,
  bestfits,
  exp.by = 0.05,
  vshift = 2,
  hshift = 0.002,
  sort.rows = F
)
}
\arguments{
\item{xmat}{A matrix of spectral data with rows corresponding to spectra and columns to ppm values}

\item{ppm}{A vector of ppm values corresponding to the columns of \code{xmat}}

\item{label}{A character string indicating the feature label to be used in the plot title}

\item{bestfits}{A list containing the best fit data for each feature, as generated by \code{\link{project_features.getBestFits}}}

\item{exp.by}{The amount by which to expand the range of each feature's ppm range in the plot}

\item{vshift}{The amount by which to vertically shift each spectrum in the plot}

\item{hshift}{The amount by which to horizontally shift each spectrum in the plot}

\item{sort.rows}{A logical indicating whether or not to sort the rows of the spectra in each plot by their summed intensity}
}
\value{
A list containing the plot objects and parameters used to create the plot
}
\description{
Generate a stacked area plot of the spectra for each feature in \code{xmat}, overlaid with the corresponding best-fit features
}
