% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prominences.R
\name{prominences}
\alias{prominences}
\title{Calculate the prominence of each peak}
\usage{
prominences(peaks, v2)
}
\arguments{
\item{peaks}{A list of peak information as output from the \code{peakdet} function.}

\item{v2}{A vector of the y-values of the peaks.}
}
\value{
A vector of the prominences of each peak.
}
\description{
The prominence of a peak is the height of the peak relative to the lowest contour line that encloses
the peak and no higher peak. This function calculates the prominence of each peak in a vector.
}
