% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pw_lags_relative_to.R
\name{pw_lags_relative_to}
\alias{pw_lags_relative_to}
\title{Since pw feature alignments can have the feauture of interest in the f1 or f2
column, provide a function which formats lags relative to the specified feature.
Note: Reversing lags also means negating them.}
\usage{
pw_lags_relative_to(lag.table, feat.num)
}
\arguments{
\item{lag.table}{lag table from feat_align, etc.}

\item{feat.num}{feature number to arrange the lag table relative to}
}
\value{
subsetted and rearranged lag table
}
\description{
Since pw feature alignments can have the feauture of interest in the f1 or f2
column, provide a function which formats lags relative to the specified feature.
Note: Reversing lags also means negating them.
}
