% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.intersect.R
\name{range.intersect.all}
\alias{range.intersect.all}
\title{Calculate pairwise intersection or union between ranges}
\usage{
\method{range}{intersect.all}(ranges, operation = "intersection")
}
\arguments{
\item{ranges}{a matrix with dimensions (n x m), where n is the number of features
and m is the number of ranges}

\item{operation}{a string indicating the type of operation to perform. Can be either
"intersection" or "union". Defaults to "intersection".}
}
\value{
a matrix with dimensions (ncol(ranges), ncol(ranges)) containing the pairwise
intersection or union between each range.
}
\description{
Given a matrix of ranges (rows: features, columns: ranges),
calculates the pairwise intersection or union between each range.
Returns a matrix with dimensions (ncol(ranges), ncol(ranges)).
}
