% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_groups.R
\name{range_groups}
\alias{range_groups}
\title{Divide Ranges into Groups with Overlapping Sections}
\usage{
range_groups(ranges, operation = "intersection", clust.by = "comps")
}
\arguments{
\item{ranges}{A matrix with two rows, representing ranges in the form c(min,max).}

\item{operation}{The type of intersection operation to use when determining overlap.}
}
\value{
A list of groups, where each group is a vector of indices
  corresponding to overlapping ranges.
}
\description{
Given a set of ranges, return a list of groups where each group contains
ranges with overlapping sections.
}
