% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range_intersect.R
\name{range_intersect}
\alias{range_intersect}
\title{Calculate the intersection or union of two numeric ranges}
\usage{
range_intersect(a, b, operation = "intersection")
}
\arguments{
\item{a}{a numeric vector representing a range}

\item{b}{a numeric vector representing a range}

\item{operation}{a character string indicating whether to calculate the "intersection" or "union" of the two ranges}
}
\value{
a numeric vector representing the intersection or union of the two ranges, or a vector of NA if the ranges do not overlap
}
\description{
Calculate the intersection or union of two numeric ranges
}
\examples{
range_intersect(c(1, 5), c(3, 6), operation = "intersection")
range_intersect(c(1, 5), c(3, 6), operation = "union")

}
