% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rebuild_specFit.R
\name{rebuild_specFit}
\alias{rebuild_specFit}
\title{...}
\usage{
rebuild_specFit(sf, feature, xmat, refmat)
}
\arguments{
\item{sf}{spec feature information (includes backfit information)}

\item{feature}{features object for study}

\item{xmat}{spectral matrix for study}

\item{refmat}{reference spectral matrix for study (spectra on rows)}
}
\value{
just the fit data
}
\description{
...
}
