% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_spectra.R
\name{resample_spectra}
\alias{resample_spectra}
\title{resample_spectra}
\usage{
resample_spectra(xmat, ppm, npoints, cores = 1)
}
\arguments{
\item{xmat}{spectral matrix (spectra on rows)}

\item{ppm}{ppm vector to match xmat columns}

\item{npoints}{number of points to include}

\item{cores}{number of cores for parallel processing (default 1)}
}
\value{
spectral matrix and ppm value linearly interpolated to new number of points
}
\description{
Linearly interpolates (or decimates) a ppm vector and spectral matrix to a 
(typically lower) number of spectral points.
Preserves spectral range and intensities. Should be replaced with binning or 
moving average at some point.
}
