% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rm.covered.points.R
\name{rm.covered.points}
\alias{rm.covered.points}
\title{Remove points on each line which are covered by the AUC of lower rows (to give
the effect of spectra covering each other up when rastering instead of plotting
multiple geom_area curves in a loop).}
\usage{
rm.covered.points(mat, feats = NULL, ss.rows = NULL)
}
\arguments{
\item{mat}{xmat subset (row and col) that will be plotted}

\item{feats}{features lined up with cols of mat. Corresponding row in mat (for each feature) is given in ss.rows.}

\item{ss.rows}{row in mat for each feature (row in feats) - often duplicates in this vector}
}
\value{
filter for the features matrix
}
\description{
If only mat (spectral matrix), then calc for mat.
}
\details{
If feats and ss.rows are supplied, assume we want to remove feat points that 
are covered by mat. Return feats matrix after filtering (NA addition)
ss.rows are the row inds in mat corresponding to each row of feats.
length(ss.rows) == nrow(feats), and all(unique(ss.rows) %in% 1:nrow(mat))
}
