% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_labels.R
\name{run.labels}
\alias{run.labels}
\title{Assigns a unique label to each run of 1s in a binary vector.}
\usage{
run.labels(v)
}
\arguments{
\item{v}{A binary vector.}
}
\value{
A numeric vector with the same length as `v` containing the labels
  assigned to each run of 1s.
}
\description{
This function takes a binary vector and assigns a unique integer label to each
run of 1s in the vector. The labels start at 1 and increment by 1 for each new
run. Runs of 0s are assigned a label of 0.
}
\examples{
v <- c(0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1)
run.labels(v)
 
}
