% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tina_featurespace.shiny.R
\name{runFeatureSpace.shiny}
\alias{runFeatureSpace.shiny}
\title{Run Feature Space Shiny app}
\usage{
runFeatureSpace.shiny(study, umap.obj, cluster.labs, featureStack)
}
\arguments{
\item{study}{A character string specifying the accession number of the study.}

\item{umap.obj}{A UMAP object, output of umap() or a related function.}

\item{cluster.labs}{A vector of cluster assignments for each point in the UMAP layout.}

\item{featureStack}{A object of extracted features for each point in the UMAP layout.}
}
\description{
Runs a Shiny app that displays a UMAP layout with associated clusters, and allows the user to select
points and view their associated extracted features.
}
\author{
MTJ2022
}
