% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_labels.R
\name{runs.labelBy.lengths}
\alias{runs.labelBy.lengths}
\title{Computes the lengths of runs of 1s in a binary vector.}
\usage{
runs.labelBy.lengths(v)
}
\arguments{
\item{v}{A binary vector.}
}
\value{
A numeric vector containing the lengths of all runs of consecutive
  1s in the input vector.
}
\description{
This function takes a binary vector and computes the lengths of all runs of
consecutive 1s in the vector. The output is a numeric vector, with length equal
to the number of 1s in the input vector, where each point in each run is labeled 
by its run length.
}
\examples{
v <- c(0, 1, 1, 0, 1, 1, 1, 1, 0, 1, 1, 1, 0, 0, 1, 1, 1, 1)
runs.labelBy.lengths(v)

}
