% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale.between.R
\name{scale.between}
\alias{scale.between}
\title{Scale values in a vector to be between a specified range.}
\usage{
\method{scale}{between}(data, lower = 0, upper = 1)
}
\arguments{
\item{data}{A numeric vector to be scaled.}

\item{lower}{A numeric value representing the lower bound of the output range. Default is 0.}

\item{upper}{A numeric value representing the upper bound of the output range. Default is 1.}
}
\value{
A numeric vector with values scaled between the lower and upper bounds.
}
\description{
This function takes a numeric vector and scales the values to be between a specified lower and upper bound.
}
\examples{
x <- c(1, 3, 5, 7, 9)
scale.between(x, 0, 10)

}
