% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.wasserstein.R
\name{score.wasserstein}
\alias{score.wasserstein}
\title{Calculate the Wasserstein distance between two vectors using transport::wasserstein1d.}
\usage{
score.wasserstein(v1, v2)
}
\arguments{
\item{v1}{numeric vector of values.}

\item{v2}{numeric vector of values.}
}
\value{
The Wasserstein distance between v1 and v2.
}
\description{
Calculate the Wasserstein distance between two vectors using transport::wasserstein1d.
}
\examples{
v1 <- c(1,2,3,4,5)
v2 <- c(3,5,7,9,11)
score.wasserstein(v1,v2)

}
