% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.evidence_refmet.R
\name{select.evidence_refmet}
\alias{select.evidence_refmet}
\title{Filters for relevant ref, match.info, and ref-feature info depending on 
selected from the resultsViewer panes. Returns a list that can be used for 
project_features.stackplot.
Filters evidence based on:
ref (row in heatmap)
columns (samples in scores scatterplot)
ref region (ppm of selection box in ref spectrum plot) 
ppm.tolerance between ref feats and spec, and ref feats and selected ref region
cutoff.residuals.feat 
cutoff.residuals.spec}
\usage{
select.evidence_refmet(
  ref.ind = 1,
  spec.ind = NULL,
  match.info,
  backfits,
  rfs.used,
  lib.data.processed,
  xmat,
  ppm,
  ppm.tolerance = 0.2,
  cutoff.residuals.feat = 0.5,
  cutoff.residuals.spec = 0.5
)
}
\arguments{
\item{ref.ind}{index of the ref in the ss-ref scores mat}

\item{spec.ind}{spectra indices, from scatterplot of scores for selected ref}

\item{match.info}{match.info table used for subsetting}

\item{backfits}{ref feature fit information}

\item{rfs.used}{lists of which ref features were used in each score, allows only best evidence}

\item{lib.data.processed}{dataset-interpolated ref library. contains compound names + data + more}

\item{xmat}{spectral matrix}

\item{ppm}{ppm for xmat}

\item{ppm.tolerance}{maximum allowed distance (ppm) between ref feature range in subset spectra and the selected ref region.}

\item{cutoff.residuals.feat}{currently not used}

\item{cutoff.residuals.spec}{currently not used}
}
\value{
list of evidence for given compound match to ref signature, limited to specific samples and ref spectrum region;
- ref.ind: index of ref spectrum
- ref.info: library info for ref spectrum
- rf.fits: list of fit objects for relevant ref features
}
\description{
Filters for relevant ref, match.info, and ref-feature info depending on 
selected from the resultsViewer panes. Returns a list that can be used for 
project_features.stackplot.
Filters evidence based on:
ref (row in heatmap)
columns (samples in scores scatterplot)
ref region (ppm of selection box in ref spectrum plot) 
ppm.tolerance between ref feats and spec, and ref feats and selected ref region
cutoff.residuals.feat 
cutoff.residuals.spec
}
\examples{
v1 <- c(1, 2, 3, 4, 5)
v2 <- c(3, 5, 7, 9, 11)
score.wasserstein(v1, v2)

}
