% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_evidence_refmet.R
\name{select_evidence_refmet}
\alias{select_evidence_refmet}
\title{For use in filtering evidence down to the relevant plot region and reference 
compound.}
\usage{
select_evidence_refmet(
  ref = NULL,
  sample = NULL,
  features.c,
  match.info,
  backfits,
  rfs.used,
  lib.data.processed,
  score.name = "fsaxrval",
  ppm.tolerance = pars$matching$filtering$ppm.tol,
  cutoff.residuals.feat = 0.5,
  cutoff.residuals.spec = 0.5
)
}
\arguments{
\item{ref}{ref info row (as used in resultsViewer)}

\item{sample}{sample info row (as used in resultsViewer)}

\item{match.info}{dataframe of match info (i.e., ref features)}

\item{backfits}{spectrum-level backfit info dataframe for each rf (row of  match.info)}

\item{rfs.used}{list of rfs used in each backfit (indices in match.info)}

\item{lib.data.processed}{spectral matrix- interpolated reference data}

\item{ppm.tolerance}{max allowed distance for selected ref region and spec region}

\item{cutoff.residuals.feat}{unused}

\item{cutoff.residuals.spec}{unused}
}
\value{
list containing expanded fit data for each backfit passing the filters
}
\description{
Filters for relevant ref and 
- ref list
- backfit list
- match.info table
}
\details{
The idea here is to figure out which rfs, of those used for scoring, apply
to the ref and subset spectra selected by ref.ind and spec.ind, respectively.
These rfs can then be passed onto the plotting function, e.g. fastStack.withFeatures()
}
