% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{select_features}
\alias{select_features}
\title{Compressed features object subsetting}
\usage{
select_features(compressed.feature, row.nums = NULL)
}
\arguments{
\item{compressed.feature}{compressed feature obj}

\item{just.rows}{row numbers of feature selection (or reordered row numbers)}
}
\value{
compressed matrix (list format)
         - pos : linear index within stack
         - vals: non-NA vals within stack
         - m : number of rows in stack
         - n : number of columns in stack
}
\description{
Pull selected features from compressed obj.
}
\details{
just.rows can be used to extract a subset of features or all, just in a different order.
cstack <- compress_stack()
Example: 
# Make example feature obj
 feature <- list(profile = rbind(c(NA, NA, 1, 3, 5, 3, NA, 1, NA), c(NA, 1, 3, 5, NA, 3, NA, 1, 3), c(3, 5, NA, NA, 1, NA, NA, 1, 3)))
 feature$position <- outer(1:3, 1:9, "+")
 feature$position[is.na(feature$profile)] <- NA
# Compress
 compressed.feature <- compress_features(feature)
# Decompress all
 feature.exp <- expand_features(compressed.feature)
# Decompress rows 1 & 2
 feature.exp <- expand_features(compressed.feature, c(1,2))
}
