% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{Setup function for initializing the analysis environment}
\usage{
setup(params_obj)
}
\arguments{
\item{params_obj}{An object containing various input parameters, such as the study and library directories.}
}
\value{
None
}
\description{
This function performs various setup tasks, such as copying library data and metadata, and finding the best fit for the spectrometer frequency.
}
\examples{
setup(params_obj)

}
\keyword{data}
\keyword{initialization,}
\keyword{setup,}
