% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftedInds_toVals.R
\name{shiftedInds_toVals}
\alias{shiftedInds_toVals}
\title{Extract values from a matrix given shifted indices}
\usage{
shiftedInds_toVals(xmat, shiftedInds)
}
\arguments{
\item{xmat}{A matrix containing the values to extract}

\item{shiftedInds}{A matrix of shifted indices (with the same number of rows as xmat)
indicating which values to extract from xmat}
}
\value{
A matrix containing the values of xmat at the shifted indices given in shiftedInds
}
\description{
Given a matrix and a matrix of shifted indices (with the same number of rows), extract
the values of the original matrix at those shifted indices. The shifted indices can be
created by aligning a subregion of the original matrix with a reference vector.
}
\examples{
xmat <- matrix(1:9, nrow = 3)
shiftedInds <- matrix(c(1, 2, 3, 2, 3, 4, 3, 4, 5), nrow = 3)
shiftedInds_toVals(xmat, shiftedInds)

# Output:
#      [,1] [,2] [,3]
# [1,]    1    2    3
# [2,]    2    3    4
# [3,]    3    4    5
}
