% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resultsViewer_heatmap.v3.R
\name{show.me.the.evidence}
\alias{show.me.the.evidence}
\title{show.me.the.evidence : results viewer shiny app}
\usage{
show.me.the.evidence(results.dir = NULL)
}
\arguments{
\item{results.dir}{A list of input parameters.}
}
\value{
spins up a shiny app from the results directory
}
\description{
Browse reference spectra (PRCSs) and metabolite-sample summed scores, and drill
down to the evidence for their associations at the multiple levels.
}
\details{
To use, call with the results directory.
}
