% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_me_the_evidence.R
\name{show_me_the_evidence}
\alias{show_me_the_evidence}
\title{show_me_the_evidence : results viewer shiny app}
\usage{
show_me_the_evidence(results.dir = NULL)
}
\arguments{
\item{results.dir}{A list of input parameters.}
}
\value{
spins up a shiny app from the results directory
}
\description{
Browse reference spectra (PRCSs) and metabolite-sample summed scores, and drill
down to the evidence for their associations at the multiple levels.
}
\details{
To use, call with the results directory.
}
