% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simplePlot.R
\name{simplePlot}
\alias{simplePlot}
\title{Create a simple plot using ggplot2}
\usage{
simplePlot(
  ymat = NULL,
  xvect = NULL,
  n_xticks = NULL,
  xdir = "reverse",
  linecolor = "gray",
  opacity = 0.6,
  linewidth = 0.5
)
}
\arguments{
\item{ymat}{A matrix or dataframe, with each column being a spectrum to plot.}

\item{xvect}{A numeric vector for the x-axis values.}

\item{n_xticks}{An integer specifying the number of x-axis ticks to generate.}

\item{xdir}{A character string indicating the direction of the x-axis, either "reverse" or "forward".}

\item{linecolor}{A character string indicating the color of the plotted lines. If defining color for each line, pass a vector of strings.}

\item{opacity}{A numeric value specifying the opacity of the plotted lines.}

\item{linewidth}{A numeric value specifying the width of the plotted lines.}
}
\value{
A ggplot object.
}
\description{
This function creates a simple plot using ggplot2, based on a matrix or dataframe input.
It can handle both single vector and multiple vectors input, and can generate x-axis ticks
with a specified number of ticks or using a default breaks_pretty() function.
}
\examples{
data(mtcars)
plotmat <- scale(mtcars[, -1], center = TRUE, scale = TRUE)
simplePlot(plotmat, linecolor = "black", opacity = 0.6, linewidth = 0.5)

simplePlot(plotmat[, 1], linecolor = "red", opacity = 0.6, linewidth = 0.5)

simplePlot(plotmat, xvect = mtcars$wt, n_xticks = 5, linecolor = "black", opacity = 0.6, linewidth = 0.5)

}
