% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortedStack.R
\name{sortedStack}
\alias{sortedStack}
\title{sortedStack}
\usage{
sortedStack(mat, ppm = NULL, vshift = 10, hshift = 0)
}
\arguments{
\item{mat}{The matrix to sort and stack.}

\item{ppm}{The x-axis values for the columns of the matrix. Defaults to NULL.}

\item{vshift}{The vertical shift for the stacked plots. Defaults to 10.}

\item{hshift}{The horizontal shift for the stacked plots. Defaults to 0.}
}
\value{
A stacked plot of the sorted matrix.
}
\description{
Sorts a matrix by row sum and stacks the rows.
}
