% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/span.R
\name{span}
\alias{span}
\title{Calculates the span of a numeric vector}
\usage{
span(v)
}
\arguments{
\item{v}{A numeric vector}
}
\value{
The span of the vector
'
}
\description{
Calculates the span of a numeric vector, which is defined as the range of the values
plus one.
}
