% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stackplot.R
\name{stackplot}
\alias{stackplot}
\title{Plot stacked spectra with ggplot2 and ggridges}
\usage{
stackplot(
  ymat = NULL,
  xvect = NULL,
  vshift = 1,
  hshift = 0,
  shift_by = NULL,
  xdir = "reverse",
  show = TRUE
)
}
\arguments{
\item{ymat}{A numeric matrix of spectral intensities.}

\item{xvect}{A numeric vector of ppm values.}

\item{vshift}{A numeric value for shifting spectra vertically (spectra are scaled down to this value).}

\item{hshift}{A numeric value for shifting spectra horizontally (as % of actual range).}

\item{shift_by}{A numeric vector for shifting spectra by row number.}

\item{xdir}{A character string for x-axis direction, either "reverse" or "forward".}

\item{show}{A logical value for whether or not to display the plot.}
}
\value{
A ggplot2 object.
}
\description{
This function plots stacked spectra using ggplot2 and ggridges, which
creates a density plot of the spectra that can help to visualize overlaps
and peaks.
}
\details{
Note: this uses specific settings for the x-axis: breaks=breaks_pretty(), limits = xlim_fixed, expand = c(0, 0), oob = scales::oob_keep
}
\examples{
data(metab)
stackplot(metab[, 1:10], xvect = metab[, 11])

}
