% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_decomp.R
\name{stat_decomp}
\alias{stat_decomp}
\title{Statistical Decomposition for spectra.}
\usage{
stat_decomp(peaks, spec, params)
}
\arguments{
\item{peaks}{picked peaks for representative spectrum, numeric vector
containing ppm indices of peaks}

\item{spec}{spectral matrix consisting of n columns (one for each ppm
 value). row  1 (named 'ppm') is the vector of ppm values. rows 2:m+1
(names undefined) are the m sample measurements for each ppm value.}

\item{params}{yaml object containg parameters for the pipeline.}
}
\value{
list of:
  target.RDS: list with one element for each cluster, containing:
      [1] ppm vector
      [2] thresholded intensities for each ppm value
  ppeaks.RDS: list with picked peaks from each cluster.
  One element per cluster, containing:
      numeric vector
        col 1: (named "chemical-shift") ppm values
        col 2: (named "intensity") intensities (covariance from STOCSY)
}
\description{
Given a spectral matrix and a list of picked peaks, loop
through each picked peak and perform a STOCSY using it as
a driver peak. Threshold the result at r >= 0.8, giving
pseudospectra (target list). Pick peak these using standard settings,
and store the peak clusters in ppeaks.
}
