% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocsy_slim.R
\name{stocsy_slim}
\alias{stocsy_slim}
\title{Calculate Pearson correlations and covariance using STOCSY for a subset of variables.}
\usage{
stocsy_slim(X, driverInd)
}
\arguments{
\item{X}{A matrix of spectral data with variables in the columns.}

\item{driverInd}{A vector of indices for the variables that will be used as drivers.}
}
\value{
A list with two elements: \code{cc}, a vector of Pearson correlations 
        between each variable in \code{X} and the variables in \code{driverInd}; and
        \code{cv}, a vector of covariance values between each variable in \code{X} 
        and the variables in \code{driverInd}.
}
\description{
Calculate Pearson correlations and covariance using STOCSY for a subset of variables.
}
