% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stocsydec.R
\name{stocsydec}
\alias{stocsydec}
\title{stocsydec}
\usage{
stocsydec(spectra, idx, cthr = 0.8)
}
\arguments{
\item{spectra}{individual spectra file to run through decomp}

\item{idx}{index number of spextra files to be run through decomp}

\item{cthr}{cutoff threshold}
}
\value{
Result of STOCS decomposition.
}
\description{
Function to run 'STOCSY decomposition' Goncalo Graca, 15 September 2020, g.gomes-da-graca/at/imperial.ac.uk
}
