% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub2indR.R
\name{sub2indR}
\alias{sub2indR}
\title{sub2indR}
\usage{
sub2indR(rows, cols, m)
}
\arguments{
\item{rows}{rows object}

\item{cols}{columns object}

\item{m}{coefficient}
}
\value{
linear indices
}
\description{
Convert row, column subscripts to linear indices.
Comparable to sub2ind in MATLAB.
}
\details{
MTJ mjudge-at-imperial.ac.uk
9MARCH2022
}
