% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair_score_summation.R
\name{sum_score}
\alias{sum_score}
\title{Sum the score in the scoresObj}
\usage{
sum_score(score.obj, refspec)
}
\arguments{
\item{score.obj}{list containing
- score.name (must be a column name in rp), 
- score.vect (vector of length(refspec) indexed by ref.pts) containing current scores for each ref point
- rf.vect (vector of length(refspec) indexed by ref.pts) containing corresponding rf.id for rf which set the current score at each point}

\item{rp}{ref.pair row}

\item{ref.pts}{indices of score.vect and rf.vect corresponding to this rf; those we are currently updating}
}
\value{
score object with relevant points updated in both vectors  
Example: bff.res <- update_scoreObj(ref.pair[j, ], bff.res, ref.pts)
}
\description{
Sum the score in the scoresObj
}
