% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_clean_nmr.R
\name{theme_clean_nmr}
\alias{theme_clean_nmr}
\title{ggplot theme for viewing nmr spectra}
\usage{
theme_clean_nmr()
}
\value{
changes theme on current plot
}
\description{
all white background, few things
}
