% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tina.R
\name{tina}
\alias{tina}
\title{Tina Function. Tina is not annotation.}
\usage{
tina(pars)
}
\arguments{
\item{pars}{A list of parameters for the TINA pipeline.}
}
\value{
A list containing the results and cluster labels from the TINA pipeline.
}
\description{
Function for running TINA stage of pipe pipeline.
}
