% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tina_umap_apcluster.R
\name{tina_combineFeatures}
\alias{tina_combineFeatures}
\title{Combines and clusters features in a matrix}
\usage{
tina_combineFeatures(
  featureStack,
  doUMAP = TRUE,
  umap.n_neighbors = 20,
  umap.min_dist = 0.1,
  umap.n_epochs = 500,
  ap.rcutoff = 0.99,
  ap.q = 0.95,
  ap.max.iter = 1000,
  ap.lam = 0.9,
  max.plots = 250,
  plot.loc = "./",
  plot.name = "study_feature_clusters.pdf"
)
}
\arguments{
\item{featureStack}{A matrix of features to be clustered.}

\item{doUMAP}{Logical indicating whether to perform UMAP projection on the features.}

\item{umap.n_neighbors}{the number of approximate nearest neighbors used to construct 
the initial high-dimensional graph. It effectively controls how UMAP balances local 
versus global structure - low values will push UMAP to focus more on local structure by 
constraining the number of neighboring points considered when analyzing the data in high dimensions, 
while high values will push UMAP towards representing the big-picture structure while losing fine detail.}

\item{umap.min_dist}{the minimum distance between points in low-dimensional space. This parameter controls
 how tightly UMAP clumps points together, with low values leading to more tightly packed embeddings. 
Larger values of min_dist will make UMAP pack points together more loosely, focusing instead on the 
preservation of the broad topological structure.}

\item{umap.n_epochs}{The number of iterations to run.}

\item{ap.rcutoff}{rcutoff for filtering out correlations before clustering}

\item{ap.q}{q parameter is the shared value for the initial preferences s(i,k) for a priori clustering. See apcluster doc.}

\item{ap.max.iter}{The maximum number of iterations for the ap cluster.}

\item{ap.lam}{The "self-responsibility" parameter for the affinity propagation clustering algorithm.}

\item{max.plots}{The maximum number of scatter plots to generate for the clusters.}

\item{plot.loc}{The directory in which to save the scatter plot PDF.}

\item{plot.name}{The name of the scatter plot PDF.}
}
\value{
A list containing UMAP and clustering results, cluster assignments, pairwise
distances and correlations.
}
\description{
Combines and clusters features in a matrix
}
