% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tina_setup.R
\name{tina_setup}
\alias{tina_setup}
\title{TINA Setup}
\usage{
tina_setup(a, xmat)
}
\arguments{
\item{a}{A list of peak regions.}

\item{xmat}{A matrix of spectra.}
}
\value{
A list containing the feature stack, position stack, subsets, and 
region information.
}
\description{
Pre-compute a bunch of useful stuff for TINA and store in an organized way ("feature" object)
Given a list of peak regions and a matrix of spectra, creates a matrix to house
all the features and returns it in a list format along with additional 
information about the subsets and regions of interest.
}
\examples{
# Create a mock input list of peak regions and a matrix of spectra
peaks <- list(list(finalRegion = c(1, 2, 3), ref.vals = c(1, 2, 3), 
                   ref.idx = c(1, 2, 3), subset = c(1, 2)),
              list(finalRegion = c(4, 5, 6), ref.vals = c(4, 5, 6), 
                   ref.idx = c(4, 5, 6), subset = c(3, 4)))
spectra <- matrix(1:12, ncol = 4)

# Run the function
tina_setup(a = peaks, xmat = spectra)
}
