% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trim_sides.R
\name{trim_sides}
\alias{trim_sides}
\title{Trim the sides of a matrix to remove columns with all NA values}
\usage{
trim_sides(mat, out = "mat")
}
\arguments{
\item{mat}{a matrix or data.frame to be trimmed}

\item{out}{character string indicating the output type. Possible values are "mat"
(default), which returns the trimmed matrix, or "inds", which returns the indices
of the columns that were kept.}
}
\value{
a matrix with the same number of rows as \code{mat} but fewer columns,
with columns containing NA values taken out.
}
\description{
Trim the sides of a matrix to remove columns with all NA values
}
\examples{
mat <- matrix(c(1,2,NA,4,NA,NA,7,8,NA), nrow = 3)
trim_sides(mat)

}
