% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_indexing.R
\name{unzip_studies}
\alias{unzip_studies}
\title{Unzips studies found in a subdirectory}
\usage{
unzip_studies(data.dir, exclude = NULL)
}
\arguments{
\item{data.dir}{directory containing zip files of your run results Other files and unzipped results can be in there, but any zip file will be assumed to be a run result}

\item{exclude}{(optional) list of run names that are problematic (i.e. cannot be unzipped correctly for some reason, or which you don't want to unzip)}
}
\value{
nothing
}
\description{
There are several operations that need to be done for SAFER results to be used
(i.e.) with browse_evidence().
}
\details{
Results are written as zip files for portability. 

The idea is that you have a local location where you keep your SAFER results, 
and re-run this function to unpack any new results. 

This command
 - unzips files which haven't previously been unzipped
 - excludes any studies you explicitly name
 - checks the heatmap html file to ensure it was written correctly, and 
   corrects it if needed

Examples:
 data.dir <- '/Users/mjudge/Documents/safer/results/' # this is where you put your zip files
 unzip_studies(data.dir) 
 unzip_studies(data.dir, exclude = c('1697751863','1697759923')) # you can also exclude problematic runs
}
