% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/valid_pars.R
\name{valid_pars}
\alias{valid_pars}
\title{Parameter validation (using params object)}
\usage{
valid_pars(pars)
}
\arguments{
\item{pars}{A list of input parameters.}
}
\value{
Adjusted parameters
}
\description{
Correct any obviously incorrect params. Replace any empties with default params.
}
\details{
Then go through a bunch of checks to see if params are reasonable.
}
