% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vectInds.R
\name{vectInds}
\alias{vectInds}
\title{Return the index/indices in a vector closest to given value(s).}
\usage{
vectInds(vals, vect)
}
\arguments{
\item{vals}{a numeric vector of values to find the closest index/indices to.}

\item{vect}{a numeric vector of values to search for closest index/indices.}
}
\value{
a numeric vector of indices in vect which are closest to each value in vals.
}
\description{
Given a vector of values, find the index/indices in another vector which is closest to each value.
}
\examples{
vectInds(1:5, c(3.5, 4.6, 2.1))

}
