% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xcorr_localign.R
\name{xcorr_localign}
\alias{xcorr_localign}
\title{Cross-correlation with local alignment}
\usage{
xcorr_localign(
  x,
  ppm,
  signature.vals,
  signature.idx.wind,
  currentInds,
  min.overlap = 3,
  slide = "internal",
  lag.limit = NA,
  plots = F
)
}
\arguments{
\item{x}{A matrix where each row is a spectrum.}

\item{ppm}{A vector of column names for the matrix x.}

\item{signature.vals}{A vector of values matching the signature index window.}

\item{signature.idx.wind}{The reference indices (across all rows of currentInds).}

\item{currentInds}{The window(s) in which the reference sits in x (often larger than the reference).}

\item{min.overlap}{The minimum number of points on which to base the correlation.}

\item{slide}{If "internal", the function just slides the reference vector until it hits the bounds of currentInds;
if "external", the function pads currentInds to ensure that all reference vector points cross the bounds of currentInds;
if "limited", the user provides the number of lags (duplicated for both directions).}

\item{lag.limit}{The limit on the number of lags to consider (for "limited" slide option).}

\item{plots}{A logical value indicating whether to generate a plot showing the aligned spectra and reference vector.}
}
\value{
A list with the following elements:
\item{best.shifts}{A vector of the best shift for each spectrum.}
\item{best.rvals}{A vector of the correlation scores for each spectrum.}
\item{best.pvals}{A vector of the p-values for each spectrum.}
\item{shiftedInds}{A matrix of the indices in x for the aligned spectra.}
\item{shiftedSpecs}{A matrix of the aligned spectra.}
\item{ref.aligned}{A vector of the reference vector values (aligned with the spectra).}
\item{overlap.sizes}{A vector of the overlap sizes for each spectrum.}
\item{g}{A plot showing the aligned spectra and reference vector (if \code{plots = TRUE}).}
}
\description{
This function calculates the cross-correlation between a reference vector and a matrix of spectra, where each
spectrum is a row of the matrix. The function aligns the reference vector with each spectrum by shifting
the spectrum and then calculating the cross-correlation between the aligned spectrum and the reference vector.
The function returns the best shift for each spectrum, along with the corresponding correlation score, p-value,
and overlap size. The function can also generate a plot showing the aligned spectra and reference vector.
}
\examples{
x <- matrix(rnorm(100), ncol = 10)
ppm <- seq(100)
signature.vals <- rnorm(10)
signature.idx.wind <- seq(5, 14)
currentInds <- matrix(seq(1, 100, by = 10), ncol = 2)
xcorr_localign(x, ppm, signature.vals, signature.idx.wind, currentInds, min.overlap = 3, slide = "internal", lag.limit = NA, plots = FALSE)

}
