% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backfit_rf.2.spec.R
\name{backfit_ref.feats.2.subset.specs}
\alias{backfit_ref.feats.2.subset.specs}
\title{Back-fit reference features to a subset of spectra}
\usage{
backfit_ref.feats.2.subset.specs(
  m.inds,
  fits.feature,
  match.info,
  feature,
  xmat,
  ppm,
  plots = T
)
}
\arguments{
\item{m.inds}{A vector of indices corresponding to the matches between reference features and spectra. Necessary to allow for subsetting of matches (i.e. because of filtering)}

\item{fits.feature}{A list of fitted features (to reference spectra)}

\item{match.info}{A data frame containing information about the matches between features and reference spectra}

\item{feature}{List containing feature intensities and positions}

\item{xmat}{Full spectral matrix from which the features were derived}

\item{ppm}{A vector containing the ppm axis of the spectra}

\item{plots}{A logical value indicating whether to generate plots of the fits and back-fit feasibility scores - only use for a couple of backfits at a time as plots will double the weight of the result}
}
\value{
A list containing the back-fits and back-fit feasibility scores for each spectrum in the subset,
        as well as a grid plot of the fits and back-fit feasibility scores.
}
\description{
Back-fit reference features to a subset of spectra
}
