% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkClusters.R
\name{checkClusters}
\alias{checkClusters}
\title{Calculate feature cluster information - wrapper for check.cluster that parallelizes it
See check.cluster documentation.}
\usage{
checkClusters(clusters, feature, par.cores = 1, par.type = "PSOCK")
}
\arguments{
\item{clusters}{cluster data from tina_combineFeatures_optics (and adjustments in tina)}

\item{feature}{feature object containing data on all the features}
}
\value{
clust.info.split (list of cluster information for each cluster, concatenated)
}
\description{
Calculate feature cluster information - wrapper for check.cluster that parallelizes it
See check.cluster documentation.
}
