% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{compress_pos}
\alias{compress_pos}
\title{Compress positions vector:
 - transpose (assuming rows have runs)
 - compress sequential inds to ranges
 - return as data.frame}
\usage{
compress_pos(pos, m, n)
}
\arguments{
\item{pos}{linear inds of vals in original stack (column-wise)}

\item{m}{nrow(stack)}

\item{n}{ncol(stack)}
}
\value{
positions in original stack, as linear indices
}
\description{
Compress positions vector:
 - transpose (assuming rows have runs)
 - compress sequential inds to ranges
 - return as data.frame
}
