% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{compress_stack}
\alias{compress_stack}
\title{Sparse (NA-gapped) matrix compression}
\usage{
compress_stack(stack)
}
\arguments{
\item{stack}{NA-laden matrix with real values you want to retain}
}
\value{
compressed matrix (list format)
         - pos : linear index within stack
         - vals: non-NA vals within stack
         - m : number of rows in stack
         - n : number of columns in stack
}
\description{
Compresses matrix to its non-NA values and records their positions
to enable perfect reconstruction on-the-fly. Intended for sparse matrices
and parallel operations as well as general memory management.
Example: c.stack <- compress_stack(rbind(c(NA, NA, 1, 3, 5, 3, NA, 1, NA), c(NA, 1, 3, 5, NA, 3, NA, 1, 3), c(3, 5, NA, NA, 1, NA, NA, 1, 3)))
}
