% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{cstack_expandRows}
\alias{cstack_expandRows}
\title{Decompress cstack object. Works on full stack and selection. Will return a 
matrix with the number of rows in the cstack obj, in their original order.}
\usage{
cstack_expandRows(cstack)
}
\arguments{
\item{cstack}{compressed matrix object}
}
\value{
compressed matrix with only the selected row data
         - pos : linear index within stack
         - vals: non-NA vals within stack
         - m : number of rows in stack
         - n : number of columns in stack
}
\description{
Decompress cstack object. Works on full stack and selection. Will return a 
matrix with the number of rows in the cstack obj, in their original order.
}
\examples{
rows.expanded <- cstack_expandRows(cstack)
          rows.expanded <- cstack_expandRows(selection)

}
