% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compression.R
\name{expand_stacklist}
\alias{expand_stacklist}
\title{Get each stack(s) from co-compressed stacklist}
\usage{
expand_stacklist(c.stack, which.stacks = NULL)
}
\arguments{
\item{c.stack}{compressed stack where vals is a list of co-compressed stacks}

\item{which.stacks}{inds or names of stacks to return from vals}
}
\value{
ranges of the incremental runs in keep

Example: all.equal(expand_pos(compress_pos(keep)), keep)
}
\description{
Get each stack(s) from co-compressed stacklist
}
