% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.matches_shiftDelta.R
\name{filter.matches_shiftDelta}
\alias{filter.matches_shiftDelta}
\title{Filter matches based on ppm shift difference}
\usage{
\method{filter}{matches_shiftDelta}(match.info, feature, ppm, ppm.tol = 0.5)
}
\arguments{
\item{match.info}{A data frame containing the match information; produced in filter.matches()}

\item{feature}{Object containing features' intensity matrix, position (column inds of spectral matrix) matrix, and other information about the features}

\item{ppm}{A numeric vector of ppm values; feature$position values index this}

\item{ppm.tol}{The maximum allowed ppm shift difference between the feature and the match}

\item{fits.feature}{List of feature fits to reference spectral regions}
}
\value{
A list with two elements: the filtered `match.info` data frame, and the filtered `fits.feature` list which have been filtered by ppm shift difference.
}
\description{
This function filters the matches in match.info based on the difference in
ppm shift between the feature and the match. Only matches with a ppm shift
difference less than or equal to ppm.tol are retained.
}
