% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_matches.R
\name{filter_matches}
\alias{filter_matches}
\title{Match Filtering}
\usage{
filter_matches(pars)
}
\arguments{
\item{pars}{A list of input parameters.}
}
\value{
A list of filtered and processed matched peak information, including back-fits to the original spectra.
}
\description{
Filter and process matched peaks based on user-specified criteria, such as singlet removal and ppm distance filtering.
Also calculate backfits of ref-feats (reference subsignatures fished out by feature matches to reference spectra) to each individual dataset spectrum.
Note: ref-feat fit is obtained using the original feature, then backfit feasibility scores are calculated. BFF scores indicate the extent to which ref-feat resonances do not exceed actual spectral signal. Specifically, each single ref-feat resonance positive residual is evaluated as a fraction of the total ref-feat height. Because the absence of a feasible fit for any single reference resonance invalidates a match, the worst-violating resonance gives the score for the whole ref-feat in a particular spectrum. Note that a ref-feat receives a BFF score for each spectrum it is fit to.
}
