% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fse.R
\name{fse}
\alias{fse}
\title{FSE Function}
\usage{
fse(pars)
}
\arguments{
\item{pars}{A list of parameters for the function.}
}
\value{
A list of features extracted from the spectra.
}
\description{
Extracts feature shapes from a matrix of NMR spectra using the FSE algorithm.
}
