% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurricane.R
\name{hurricane}
\alias{hurricane}
\title{hurricane}
\usage{
hurricane(params_loc)
}
\arguments{
\item{params_loc}{location of the params.yaml file containing
pipeline run parameters.}
}
\value{
N/A but .RDS and .TSV files are saved, see description.
}
\description{
Run the NMR pipeline start to finish:
1] statistical decomposition
2] annotation matching
3] export results
4] visualisation [in progress in upstream repository]
Also saves to RDS outputs from statistical decomposition
A default params.yaml file will be used if none is supplied.
}
