% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind2subR.R
\name{ind2subR}
\alias{ind2subR}
\title{ind2subR}
\usage{
ind2subR(inds, m)
}
\arguments{
\item{inds}{The linear indices to be converted IE linear index
(1 = [1,1], 2 = [2,1], 3 = [3,1], 4 = [1,2], 5 = [2,2])}

\item{m}{The number of rows IE m = nrow(matrix)}
}
\value{
List of row and column indices.
}
\description{
Utility function for converting linear indices (as defined in MATLAB) to row
and column indices. Got the solution from
http://cran.r-project.org/doc/contrib/Hiebeler-matlabR.pdf via
https://stackoverflow.com/questions/4452039/converting-between-matrix-subscripts-and-linear-indices-like-ind2sub-sub2ind-in
}
\details{
MTJ 28FEB2022
}
