% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.features2refs.par.explicit.R
\name{match.features2refs.par.explicit}
\alias{match.features2refs.par.explicit}
\title{match.features2refs.par.explicit}
\usage{
match.features2refs.par.explicit(pars)
}
\arguments{
\item{pars}{A list of parameters containing directories, parallel settings, and matching settings.}
}
\value{
A list of matches between features and reference spectra.
}
\description{
This function matches features to reference spectra using a parallelized loop.
}
