% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pair.score.summation.R
\name{pair.score.summation}
\alias{pair.score.summation}
\title{Calculates interaction scores between all reference spectra and spectral sets.}
\usage{
pair.score.summation(pars)
}
\arguments{
\item{pars}{A list of parameters specifying input/output directories and parallelization settings.}
}
\value{
A table of scores for all possible reference spectra and spectral sets.
}
\description{
Calculates interaction scores between all reference spectra and spectral sets.
}
