% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pipeline.R
\name{pipeline}
\alias{pipeline}
\title{Feature shape based annotation pipeline ####
 /
 To run from here, you need the following files in ./data:
 params.yaml         # parameter file
 spectral.matrix.RDS # spectral matrix
 data.list.RDS       # library files
 lib.info.RDS        # index for library files}
\usage{
pipeline(params_loc, params_obj)
}
\arguments{
\item{params_loc}{Path to a YAML file containing user-specified parameters}

\item{params_obj}{List of parameters}
}
\description{
Running the setup script will accomplish this, as well as ensuring that the correct
 library is being used.
 This pipeline accepts a spectral matrix (preferably with minimal processing; e.g.
 no strong normalization, no scaling, no alignment) in the form of and RDS file
 with a matrix (row 1 = ppm vector, each additional row is a 1D NMR sample).
}
\details{
# MTJ 2023
}
