% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_corrPocketPairs.R
\name{plot_corrPocketPairs}
\alias{plot_corrPocketPairs}
\title{Plot Stackplot of Correlation Pocket Pairs}
\usage{
plot_corrPocketPairs(
  pocketPairs,
  region = NULL,
  vshift = 5,
  xvect = NULL,
  show = TRUE
)
}
\arguments{
\item{pocketPairs}{Result of corrPocketPairs_al(...) with elements named "peakMap", "corr", "cov", and "regions".}

\item{region}{A vector of integers specifying the columns of the peakMap to be included in the plot.}

\item{vshift}{The vertical shift to apply to the distributions for stacking.}

\item{xvect}{The x values at which to plot the distributions.}

\item{show}{Logical. Whether or not to display the plot.}
}
\value{
The corrPocketPairs stack plot.
}
\description{
This function plots a stackplot of correlation pocket peaks (same data as the heatmap, but plotted like spectral regions)
}
\examples{
# Load data
pocketPairs <- corrPocketPairs_al(...)

# Plot correlation heatmaps
plot_corrPocketPairs(pocketPairs, region = 1000:1500)


}
