% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/range.dist.R
\name{range.dist}
\alias{range.dist}
\title{Calculate range distances}
\usage{
\method{range}{dist}(a, b)
}
\arguments{
\item{a}{range 1}

\item{b}{range 2}
}
\value{
distance between them
}
\description{
Given a pair of ranges, return distance between them (number of elements). Overlapping
ranges return negative distance.
}
