% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.matches.R
\name{score.matches}
\alias{score.matches}
\title{score.matches function}
\usage{
score.matches(pars)
}
\arguments{
\item{pars}{a list containing necessary parameters for the function}
}
\value{
RDS file containing match scores between library reference spectra and the best subset spectrum score for each.
}
\description{
This function computes match scores between subset spectra and library reference spectra.
Uses pair.score.summation() to actually compute the scores
It builds a ss-ref matrix and looks for any compounds that match known annotations.
}
