% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slidingCorr.R
\name{slidingCorr}
\alias{slidingCorr}
\title{Calculate sliding correlation matrix and pockets}
\usage{
slidingCorr(
  x,
  ws,
  extractPockets = FALSE,
  plotting = TRUE,
  vshift = 20,
  ppm = NULL
)
}
\arguments{
\item{x}{Input matrix.}

\item{ws}{Size of sliding window.}

\item{extractPockets}{If \code{TRUE}, calculate correlation pockets for each column of \code{x} within each sliding window.}

\item{plotting}{If \code{TRUE}, produce a stacked plot of the correlation matrix with pockets highlighted.}

\item{vshift}{Vertical shift in plot.}

\item{ppm}{X-axis values for plot. If not specified, uses column numbers of \code{x}.}
}
\value{
A list containing the following items:
\item{corr_compact}{The compact form of the sliding correlation matrix.}
\item{indsmat}{The offset matrix used to calculate the sliding correlation matrix.}
\item{cov_compact}{The compact form of the sliding covariance matrix.}
\item{isPocket}{A logical matrix indicating whether each point in the sliding window is within a correlation pocket.}
\item{plot}{If \code{plotting} is \code{TRUE}, a stacked plot of the correlation matrix with pockets highlighted.}
\item{window}{The range of the sliding window.}
\item{center}{The center position of the sliding window.}
}
\description{
Calculate sliding correlation matrix and pockets
}
\examples{
# Generate a test matrix
x <- matrix(rnorm(2000), nrow = 100)

# Calculate sliding correlation matrix with pockets and plot
slidingCorr(x, ws = 20, extractPockets = TRUE, plotting = TRUE)

}
