% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_decomp.R
\name{stat_decomp}
\alias{stat_decomp}
\title{Statistical Decomposition for spectra.}
\usage{
stat_decomp(peaks_loc, spec_loc, params)
}
\arguments{
\item{peaks_loc}{location of the rds peaks file.}

\item{spec_loc}{location of the rds spec file.}

\item{params}{location of the yaml params file.}
}
\value{
list of stuff
}
\description{
Description
}
