% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat_decomp.R
\name{stat_decomp}
\alias{stat_decomp}
\title{Statistical Decomposition for spectra.}
\usage{
stat_decomp(peaks, spec, params)
}
\arguments{
\item{params}{yaml object containg parameters for the pipeline.}

\item{peaks_loc}{rds peaks file loaded into memory.}

\item{spec_loc}{rds spec file loaded into memory.}
}
\value{
list of stuff
}
\description{
Description
}
