% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storm_mtj.R
\name{storm_mtj}
\alias{storm_mtj}
\title{My attempt at a pure STORM as described in the manuscript's SI.}
\usage{
storm_mtj(
  X = NULL,
  ppm = NULL,
  b = 30,
  q = 0.05,
  ref.index = NULL,
  driver = NULL
)
}
\arguments{
\item{X}{A matrix of spectral data (rows are spectra, columns are spectral points)}

\item{ppm}{A vector of the spectral points in ppm (optional, default is all columns of X)}

\item{b}{An integer giving the expansion parameter for the reference peak}

\item{q}{A numeric giving the p-value threshold for correlation significance (both subset AND reference optimization)}

\item{ref.index}{A vector of the spectral points (columns of X) to use as the initial reference}

\item{ref}{integer (if providing best spectrum) or vector (if providing feature shape)}
}
\value{
A list
}
\description{
My attempt at a pure STORM as described in the manuscript's SI.
}
