% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub2indR.R
\name{sub2indR}
\alias{sub2indR}
\title{Convert Subscripts to Linear Indices in Column-Major Order}
\usage{
sub2indR(rows, cols, m)
}
\arguments{
\item{rows}{The row subscripts.}

\item{cols}{The column subscripts.}

\item{m}{The number of rows in the matrix.}
}
\value{
The linear indices corresponding to the input subscripts.
}
\description{
This function converts subscripts of a matrix to their corresponding linear indices in a column-major order. Intended to replicate MATLAB's sub2ind.
}
\examples{
sub2indR(1:3, 1:3, 3)
# [1] 1 4 7 2 5 8 3 6 9

}
